% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{add_step}
\alias{add_step}
\title{Add a New Step to Current Recipe}
\usage{
add_step(rec, object)
}
\arguments{
\item{rec}{A \code{\link[=recipe]{recipe()}}.}

\item{object}{A step object.}
}
\value{
A updated \code{\link[=recipe]{recipe()}} with the new step in the last slot.
}
\description{
\code{add_step} adds a step to the last location in the recipe.
}
\concept{
preprocessing
}
\keyword{datagen}
