% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dtplyr_functions.R
\name{bind_cols_dtplyr}
\alias{bind_cols_dtplyr}
\title{Bind a dtplyr table with another table
Uses join to avoid errors when using bind_cols on dtplyr tables.
Uses bind_cols otherwise}
\usage{
bind_cols_dtplyr(
  ...,
  .name_repair = c("unique", "universal", "check_unique", "minimal")
)
}
\arguments{
\item{...}{two or more tables}

\item{.name_repair}{unique, universal, check_unique, minimal
ignored for dtplyr tables}
}
\description{
Bind a dtplyr table with another table
Uses join to avoid errors when using bind_cols on dtplyr tables.
Uses bind_cols otherwise
}
