% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dtplyr_functions.R
\name{bind_rows_dtplyr}
\alias{bind_rows_dtplyr}
\title{Bind a dtplyr table with another table by row
Uses full join to avoid errors when using bind_rows on dtplyr tables.
Uses bind_rows otherwise}
\usage{
bind_rows_dtplyr(
  ...,
  .name_repair = c("unique", "universal", "check_unique", "minimal")
)
}
\arguments{
\item{...}{two or more tables}

\item{.name_repair}{unique, universal, check_unique, minimal
ignored for dtplyr tables
importFrom data.table rbindlist as.data.table}
}
\description{
Bind a dtplyr table with another table by row
Uses full join to avoid errors when using bind_rows on dtplyr tables.
Uses bind_rows otherwise
}
