% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{biomass}
\alias{biomass}
\title{Biomass Data}
\source{
Ghugare, S. B., Tiwary, S., Elangovan, V., and Tambe, S. S. (2013).
Prediction of Higher Heating Value of Solid Biomass Fuels Using Artificial
Intelligence Formalisms. \emph{BioEnergy Research}, 1-12.
}
\value{
\item{biomass}{a data frame}
}
\description{
Ghugare et al (2014) contains a data set where different biomass fuels are
characterized by the amount of certain molecules (carbon, hydrogen, oxygen,
nitrogen, and sulfur) and the corresponding higher heating value (HHV).
These data are from their Table S.2 of the Supplementary Materials
}
\examples{
data(biomass)
str(biomass)
}
\keyword{datasets}
