% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{check_bake_cols}
\alias{check_bake_cols}
\title{Check columns are available for bake steps}
\usage{
check_bake_cols(data, col_names, step_name)
}
\arguments{
\item{data}{A tibble of the training data.}

\item{col_names}{names of columns that are checked.}

\item{step_name}{name of step this function is applied to. For error message.}
}
\description{
This internal function is to be used in the bake function to ensure that
needed columns are present in the data set.
}
\keyword{internal}
