% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colcheck.R
\name{check_cols}
\alias{check_cols}
\title{Check if all Columns are Present}
\usage{
check_cols(recipe, ..., role = NA, trained = FALSE)
}
\arguments{
\item{recipe}{A recipe object. The step will be added to the
sequence of operations for this recipe.}

\item{...}{One or more selector functions to choose which
variables are checked in the step. See \code{\link[=selections]{selections()}}
for more details. For the \code{tidy} method, these are not
currently used.}

\item{role}{Not used by this step since no new variables are
created.}

\item{trained}{A logical to indicate if the column names to check
against are gathered.}
}
\description{
\code{check_cols} creates a \emph{specification} of a recipe
step that will check if all the columns of the training frame are
present in the newdata.
}
