% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/missing.r
\name{check_missing}
\alias{check_missing}
\alias{tidy.check_missing}
\title{Check for Missing Values}
\usage{
check_missing(recipe, ..., role = NA, trained = FALSE,
  columns = NULL, skip = FALSE)

\method{tidy}{check_missing}(x, ...)
}
\arguments{
\item{recipe}{A recipe object. The check will be added to the
sequence of operations for this recipe.}

\item{...}{One or more selector functions to choose which
variables are checked in the check See \code{\link[=selections]{selections()}}
for more details. For the \code{tidy} method, these are not
currently used.}

\item{role}{Not used by this check since no new variables are
created.}

\item{trained}{A logical for whether the selectors in \code{...}
have been resolved by \code{\link[=prep]{prep()}}.}

\item{columns}{A character string of variable names that will
be populated (eventually) by the terms argument.}

\item{skip}{A logical. Should the check be skipped when the
recipe is baked by \code{\link[=bake.recipe]{bake.recipe()}}? While all operations are baked
when \code{\link[=prep.recipe]{prep.recipe()}} is run, some operations may not be able to be
conducted on new data (e.g. processing the outcome variable(s)).
Care should be taken when using \code{skip = TRUE} as it may affect
the computations for subsequent operations.}

\item{x}{A \code{check_missing} object.}
}
\value{
An updated version of \code{recipe} with the new check
added to the sequence of existing operations (if any). For the
\code{tidy} method, a tibble with columns \code{terms} (the
selectors or variables selected).
}
\description{
\code{check_missing} creates a a \emph{specification} of a recipe
operation that will check if variables contain missing values.
}
\details{
This check will break the \code{bake} function if any of the checked
columns does contain \code{NA} values. If the check passes, nothing is changed
to the data.
}
\examples{
data(credit_data)
is.na(credit_data) \%>\% colSums()

# If the test passes, newdata is returned unaltered
recipe(credit_data) \%>\%
  check_missing(Age, Expenses) \%>\%
  prep() \%>\%
  bake(credit_data)

# If your training set doesn't pass, prep() will stop with an error

\dontrun{
recipe(credit_data)  \%>\%
  check_missing(Income) \%>\%
  prep()
}

# If newdata contain missing values, the check will stop bake()

train_data <- credit_data \%>\% dplyr::filter(Income > 150)
test_data  <- credit_data \%>\% dplyr::filter(Income <= 150 | is.na(Income))

rp <- recipe(train_data) \%>\%
  check_missing(Income) \%>\%
  prep()

bake(rp, train_data)
\dontrun{
bake(rp, test_data)
}
}
