% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{check_name}
\alias{check_name}
\title{check that newly created variable names don't overlap}
\usage{
check_name(res, new_data, object, newname = NULL, names = FALSE)
}
\arguments{
\item{res}{A data frame or tibble of the newly created variables.}

\item{new_data}{A data frame or tibble passed to the bake function.}

\item{object}{A trained object passed to the bake function.}

\item{newname}{A string of variable names if prefix isn't specified
in the trained object.}

\item{names}{A logical determining if the names should be set using
the names function (TRUE) or colnames function (FALSE).}
}
\description{
\code{check_name} is to be used in the bake function to ensure that
newly created variable names don't overlap with existing names.
Throws an error if check fails.
}
\keyword{internal}
