% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{check_new_data}
\alias{check_new_data}
\title{Check for required column at bake-time}
\usage{
check_new_data(req, object, new_data)
}
\arguments{
\item{req}{A character vector of required columns.}

\item{object}{A step object.}

\item{new_data}{A tibble of data being baked.}
}
\value{
Invisible NULL. Side effects are the focus of the function.
}
\description{
When baking a step, create an information error message when a column that is
used by the step is not present in \code{new_data}.
}
\seealso{
\link{developer_functions}
}
\keyword{internal}
