% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{check_options}
\alias{check_options}
\title{Check that options argument contain the right elements}
\usage{
check_options(options, exclude = NULL, include = NULL, call = caller_env())
}
\arguments{
\item{options}{options to be checked.}

\item{exclude}{Character vector, elements that can't be present in \code{options}.}

\item{include}{Character vector, Allowed elements in \code{options}.}
}
\description{
\code{check_options()} is to be used in the prep function to ensure that \code{options}
arguments are lists and contain the right elements.
}
\seealso{
\link{developer_functions}
}
\keyword{internal}
