% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{check_type}
\alias{check_type}
\title{Quantitatively check on variables}
\usage{
check_type(dat, quant = TRUE)
}
\arguments{
\item{dat}{A data frame or tibble of the training data.}

\item{quant}{A logical indicating whether the data is expected to be numeric
(TRUE) or a factor/character (FALSE).}
}
\description{
This internal function is to be used in the prep function to ensure that
the type of the variables matches the expectation. Throws an error if
check fails.
}
\keyword{internal}
