% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dtplyr_functions.R
\name{confirm_table_format}
\alias{confirm_table_format}
\title{Helper function to address checks to convert data to a tibble.
To avoid modifying the dtplyr table, the various step functions
need to accept tibbles or dtplyr tables.
Given that these are often checks done before returning the table,
it probably makes sense to run compute on the dtplyr table.}
\usage{
confirm_table_format(x)
}
\arguments{
\item{x}{an object}
}
\value{
The object, converted to a tibble or left as a dtplyr table.
}
\description{
Helper function to address checks to convert data to a tibble.
To avoid modifying the dtplyr table, the various step functions
need to accept tibbles or dtplyr tables.
Given that these are often checks done before returning the table,
it probably makes sense to run compute on the dtplyr table.
}
