% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dtplyr_functions.R
\name{convert_named_vector_to_hierarchical_list}
\alias{convert_named_vector_to_hierarchical_list}
\title{Take a named vector with subnames indicated by "." and convert
to a list with sublists for each "."
for example, mpg.values = 1, mpg.ordered = 2, cyl.values = 3, cyl.ordered = 4
becomes list(mpg = list(values = 1, ordered = 2), cyl = list(values = 3, ordered = 4))
Used to replace things like lapply(data, get_levels)}
\usage{
convert_named_vector_to_hierarchical_list(vec, sep = ".")
}
\description{
Take a named vector with subnames indicated by "." and convert
to a list with sublists for each "."
for example, mpg.values = 1, mpg.ordered = 2, cyl.values = 3, cyl.ordered = 4
becomes list(mpg = list(values = 1, ordered = 2), cyl = list(values = 3, ordered = 4))
Used to replace things like lapply(data, get_levels)
}
