% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{covers}
\alias{covers}
\title{Raw Cover Type Data}
\source{
\url{https://archive.ics.uci.edu/ml/machine-learning-databases/covtype/covtype.info}
}
\value{
\item{covers}{a data frame}
}
\description{
These data are raw data describing different types of forest cover-types 
  from the UCI Machine Learning Database (see link below). There is one 
  column in the data that has a few difference pieces of textual 
  information (of variable lengths).
}
\examples{
data(covers)
str(covers)
}
\keyword{datasets}
