% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{credit_data}
\alias{credit_data}
\title{Credit Data}
\source{
\url{https://github.com/gastonstat/CreditScoring}, 
\url{http://bit.ly/2kkBFrk}
}
\value{
\item{credit_data}{a data frame}
}
\description{
These data are from the website of Dr. Lluís A. Belanche Muñoz by way of a 
github repository of Dr. Gaston Sanchez. One data point is a missing outcome
was removed from the original data.
}
\examples{
data(credit_data)
str(credit_data)
}
\keyword{datasets}
