% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{detect_step}
\alias{detect_step}
\title{Detect if a particular step is used in a recipe}
\usage{
detect_step(recipe, step_name)
}
\arguments{
\item{recipe}{A recipe to check.}

\item{step_name}{Character name of a step, such as \code{step_intercept}.}
}
\value{
Logical indicating if recipes contains given step.
}
\description{
Detect if a particular step is used in a recipe
}
\examples{
rec <- recipe(Species ~ ., data = iris) \%>\%
  step_intercept()

detect_step(rec, "step_intercept")
}
