% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/developer.R
\name{developer-functions}
\alias{developer-functions}
\title{Developer functions for creating recipes steps}
\description{
This page provides a comprehensive list of the functions that are exported
for creating recipes steps, and how to use them.
}
\section{Creating steps}{
\code{\link[=add_step]{add_step()}} and \code{\link[=add_check]{add_check()}} are required to be used when creating a new
step. It should be the other function call in your steps and should have the
following format:

\if{html}{\out{<div class="sourceCode r">}}\preformatted{step_example <- function(recipe,
                         ...,
                         role = NA,
                         trained = FALSE,
                         skip = FALSE,
                         id = rand_id("example")) \{
  add_step(
    recipe,
    step_example_new(
      terms = enquos(...),
      role = role,
      trained = trained,
      skip = skip,
      id = id
    )
  )
\}
}\if{html}{\out{</div>}}

\code{\link[=step]{step()}} and \code{\link[=check]{check()}} are used within the \verb{step_*_new()} function that you
use in your new step. It will be used in the following way:

\if{html}{\out{<div class="sourceCode r">}}\preformatted{step_example_new <- function(terms, role, trained, skip, id) \{
  step(
    subclass = "example",
    terms = terms,
    role = role,
    trained = trained,
    skip = skip,
    id = id
  )
\}
}\if{html}{\out{</div>}}
}

\section{Interacting with recipe objects}{
Hello
}

