% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dtplyr_functions.R
\name{dimnames.dtplyr_step}
\alias{dimnames.dtplyr_step}
\title{Add a dimnames function so that colnames works properly for dtplyr objects.}
\usage{
\method{dimnames}{dtplyr_step}(x)
}
\arguments{
\item{x}{a dtplyr_step object}
}
\value{
list with NULL for rownames, colnames of the dtplyr object.
}
\description{
Add a dimnames function so that colnames works properly for dtplyr objects.
}
