% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparsevctrs.R
\name{.recipes_preserve_sparsity}
\alias{.recipes_preserve_sparsity}
\title{Does step destroy sparsity of columns}
\usage{
.recipes_preserve_sparsity(x, ...)
}
\arguments{
\item{x}{An object.}
}
\value{
Single logical.
}
\description{
Does step destroy sparsity of columns
}
\details{
This function return \code{TRUE} if modifications by this step is expected to
destroy sparsity of its columns. It does not know whether it will happen or
not since it doesn't know if the step will select sparse columns or not.

It returns \code{FALSE} for steps that have been modified to handle sparse columns
correctly.
}
\keyword{internal}
