% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparsevctrs.R
\name{.recipes_toggle_sparse_args}
\alias{.recipes_toggle_sparse_args}
\title{Toggle all auto sparse arguments}
\usage{
.recipes_toggle_sparse_args(x, choice)
}
\arguments{
\item{x}{A recipe.}

\item{choice}{A character string for separating values.}
}
\value{
A recipe
}
\description{
Toggle all auto sparse arguments
}
\details{
If a step has an argument \code{sparse = "auto"}, then workflows can use this
function to fill these values with the preferred action. This preferred
action is calculated in workflows dependent on the model and data heuristics.
Hence why it has to be passed in.

Only arguments where \code{sparse = "auto"} are affected, thus a user can set
\code{sparse = "no"} and it will be respected.
}
\keyword{internal}
