% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dtplyr_functions.R
\name{dplyr_summarize_object}
\alias{dplyr_summarize_object}
\title{Trick to summarize a function with an arbitrary return.
For example, where the function returns an object of different sizes, like unique().
Returns a named list by column names.}
\usage{
dplyr_summarize_object(dat, cols, fn, ...)
}
\description{
Trick to summarize a function with an arbitrary return.
For example, where the function returns an object of different sizes, like unique().
Returns a named list by column names.
}
