% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dtplyr_functions.R
\name{drop_top_list_level}
\alias{drop_top_list_level}
\title{Drop the top list level, keeping everything under it
Keep the names from the top level}
\usage{
drop_top_list_level(lst)
}
\description{
Drop the top list level, keeping everything under it
Keep the names from the top level
}
