% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/profile.R
\name{fixed}
\alias{fixed}
\alias{fixed.default}
\alias{fixed.numeric}
\alias{fixed.factor}
\alias{fixed.character}
\alias{fixed.Date}
\alias{fixed.POSIXct}
\alias{prof}
\alias{prof.numeric}
\alias{prof.factor}
\alias{prof.character}
\alias{prof.Date}
\alias{prof.POSIXct}
\title{Helper Functions for Profile Data Sets}
\usage{
fixed(x, pct, index, ...)

\method{fixed}{default}(x, pct, index, ...)

\method{fixed}{numeric}(x, pct, index, ...)

\method{fixed}{factor}(x, pct, index, ...)

\method{fixed}{character}(x, pct, index, ...)

\method{fixed}{Date}(x, pct, index, ...)

\method{fixed}{POSIXct}(x, pct, index, ...)

prof(x, grid, ...)

\method{prof}{numeric}(x, grid, ...)

\method{prof}{factor}(x, grid, ...)

\method{prof}{character}(x, grid, ...)

\method{prof}{Date}(x, grid, ...)

\method{prof}{POSIXct}(x, grid, ...)
}
\arguments{
\item{x}{A vector}

\item{pct, index, ..., grid}{Options pass from \code{\link[=step_profile]{step_profile()}}}
}
\description{
Helper Functions for Profile Data Sets
}
\keyword{internal}
