% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{fully_trained}
\alias{fully_trained}
\title{Check to see if a recipe is trained/prepared}
\usage{
fully_trained(x)
}
\arguments{
\item{x}{A recipe}
}
\value{
A logical which is true if all of the recipe steps have been run
through \code{prep}. If no steps have been added to the recipe, \code{TRUE} is
returned only if the recipe has been prepped.
}
\description{
Check to see if a recipe is trained/prepared
}
\examples{
rec <- recipe(Species ~ ., data = iris) |>
  step_center(all_numeric())

rec |> fully_trained()


rec |>
  prep(training = iris) |>
  fully_trained()
}
\seealso{
\link{developer_functions}
}
