% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{get_keep_original_cols}
\alias{get_keep_original_cols}
\title{Get the \code{keep_original_cols} value of a recipe step}
\usage{
get_keep_original_cols(object)
}
\arguments{
\item{object}{A recipe step}
}
\value{
A logical to keep the original variables in the output
}
\description{
Get the \code{keep_original_cols} value of a recipe step
}
\seealso{
\link{developer_functions}
}
\keyword{internal}
