% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_types_recipes.R
\name{get_types_recipes}
\alias{get_types_recipes}
\alias{get_types_recipes.default}
\alias{get_types_recipes.character}
\alias{get_types_recipes.ordered}
\alias{get_types_recipes.factor}
\alias{get_types_recipes.integer}
\alias{get_types_recipes.numeric}
\alias{get_types_recipes.double}
\alias{get_types_recipes.Surv}
\alias{get_types_recipes.logical}
\alias{get_types_recipes.Date}
\alias{get_types_recipes.POSIXct}
\alias{get_types_recipes.list}
\alias{get_types_recipes.textrecipes_tokenlist}
\alias{get_types_recipes.hardhat_case_weights}
\title{Get types for use in recipes}
\usage{
get_types_recipes(x)

\method{get_types_recipes}{default}(x)

\method{get_types_recipes}{character}(x)

\method{get_types_recipes}{ordered}(x)

\method{get_types_recipes}{factor}(x)

\method{get_types_recipes}{integer}(x)

\method{get_types_recipes}{numeric}(x)

\method{get_types_recipes}{double}(x)

\method{get_types_recipes}{Surv}(x)

\method{get_types_recipes}{logical}(x)

\method{get_types_recipes}{Date}(x)

\method{get_types_recipes}{POSIXct}(x)

\method{get_types_recipes}{list}(x)

\method{get_types_recipes}{textrecipes_tokenlist}(x)

\method{get_types_recipes}{hardhat_case_weights}(x)
}
\arguments{
\item{x}{A character vector}
}
\description{
The \code{get_types_recipes()} generic is used internally to supply types to
columns used in recipes. These functions underlie the work that the user sees
in \link{selections}.
}
\details{
This function acts as an extended recipes-specific version of \code{\link[=class]{class()}}. By
ignoring differences in similar types ("double" and "numeric") and allowing
each element to have multiple types ("factor" returns "factor", "unordered",
and "nominal", and "character" returns "string", "unordered", and
"nominal") we are able to create more natural selectors such as
\code{\link[=all_nominal]{all_nominal()}}, \code{\link[=all_string]{all_string()}} and \code{\link[=all_integer]{all_integer()}}.
}
\examples{
\dontshow{if (rlang::is_installed("modeldata")) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}

data(Sacramento, package = "modeldata")
lapply(Sacramento, get_types_recipes)
\dontshow{\}) # examplesIf}
}
