% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selections.R
\name{has_role}
\alias{has_role}
\alias{all_predictors}
\alias{all_outcomes}
\alias{has_type}
\alias{all_numeric}
\alias{all_nominal}
\title{Role Selection}
\usage{
has_role(x = "predictor", roles = NULL)

all_predictors(roles = NULL)

all_outcomes(roles = NULL)

has_type(x = "numeric", types = NULL)

all_numeric(types = NULL)

all_nominal(types = NULL)
}
\arguments{
\item{x}{A single character string for the query.}

\item{roles}{A character string of roles for the current set of terms.}

\item{types}{A character string of roles for the current set of data types}
}
\value{
An integer vector.
}
\description{
\code{has_role}, \code{all_predictors}, and \code{all_outcomes} can be used to select variables in a formula that have certain roles. Similarly,  \code{has_type}, \code{all_numeric}, and \code{all_nominal} are used to select columns based on their data type. See \code{\link{selections}} for more details.
}
\keyword{datagen}
