% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ignore_step.R
\name{ignore_step}
\alias{ignore_step}
\title{Remove steps from recipe}
\usage{
ignore_step(x, number, id)
}
\arguments{
\item{x}{A \code{recipe} object.}

\item{number}{An integer vector, Denoting the positions of the steps that
should be removed.}

\item{id}{A character string. Denoting the \code{id} of the steps that should be
removed.}
}
\value{
a \code{recipe} object.
}
\description{
\code{ignore_step} will return a recipe without steps specified by the \code{number} or
\code{id} argument.
}
\details{
\code{number} or \code{id} must be specified. Specifying neither or both will result
in a error.
}
\examples{
rec <- recipe(mpg ~ ., data = mtcars) \%>\%
  step_dummy(all_nominal_predictors()) \%>\%
  step_impute_mean(all_numeric_predictors()) \%>\%
  step_normalize(all_numeric_predictors()) \%>\%
  step_pca(all_numeric_predictors(), id = "PCA")

ignore_step(rec, number = 1)

ignore_step(rec, number = 1:2)

ignore_step(rec, id = "PCA")
}
