% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dtplyr_functions.R
\name{is_dtplyr_table}
\alias{is_dtplyr_table}
\title{Is this a dtplyr table?}
\usage{
is_dtplyr_table(x)
}
\arguments{
\item{x}{an object}
}
\value{
TRUE if class dtplyr_step, FALSE otherwise
}
\description{
Is this a dtplyr table?
}
