% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bag_imp.R
\name{learn.bagimpute_step}
\alias{learn.bagimpute_step}
\title{Estimate the Bagged Tree Imputation Models from a Training Set.}
\usage{
\method{learn}{bagimpute_step}(x, data, ...)
}
\arguments{
\item{x}{a \code{bagimpute_step} object that contains the model objects.}

\item{data}{a tibble or data frame that contains the training set. These data will be used to compute the tree ensembles that are used when this step is applied.}

\item{...}{further arguments passed to or from other methods (not currently used).}
}
\value{
An object of class \code{bagimpute_step}.
}
\description{
For a training set of data, this function uses \code{\link[ipred]{ipredbagg}} to create models that will be used to impute missing data. This transformation only fits the models while \code{\link{process}} is used to do the imputation for specific data sets. 
#'
}
\author{
Max Kuhn
}
\concept{
preprocessing imputation
}
\keyword{datagen}

