% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/center.R
\name{learn.center_step}
\alias{learn.center_step}
\title{Estimate Means from a Training Set for Centering.}
\usage{
\method{learn}{center_step}(x, data, na.rm = TRUE, ...)
}
\arguments{
\item{x}{a \code{center_step} object that specifies which columns will be centered.}

\item{data}{A tibble or data frame that contains the training set.}

\item{na.rm}{A boolean indicates whether to remove NAs, default TRUE.}

\item{...}{further arguments passed to or from other methods (not currently used).}
}
\value{
An object of class \code{center_step}.
}
\description{
For a training set of data, this function computes the sample mean across numeric columns that require centering. Note that no data are centered by this function; see  \code{\link{process.center_step}}.
}
\author{
Max Kuhn
}
\concept{
preprocessing normalization_methods
}
\keyword{datagen}

