% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interactions.R
\name{learn.interact_step}
\alias{learn.interact_step}
\title{Estimate Interaction Information from a Training Set}
\usage{
\method{learn}{interact_step}(x, data, na.rm = TRUE, ...)
}
\arguments{
\item{x}{a \code{interact_step} object}

\item{data}{a tibble or data frame that contains the training set.}

\item{...}{further arguments passed to or from other methods (not currently used).}
}
\value{
An object of class \code{interact_step}.
}
\description{
For a training set of data, this function computes the required information to produce interaction terms.Note that no interactions are created by this function; see  \code{\link{process.interact_step}}.
}
\author{
Max Kuhn
}
\concept{
preprocessing
}
\keyword{datagen}

