% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kpca.R
\name{learn.kpca_step}
\alias{learn.kpca_step}
\title{Estimate the Kernel Principal Component Loadings from a Training Set.}
\usage{
\method{learn}{kpca_step}(x, data, ...)
}
\arguments{
\item{x}{a \code{kpca_step} object that contains the kernel PCA objects.}

\item{data}{a tibble or data frame that contains the training set. These data will be used to compute the loadings that are used when this step is applied.}

\item{...}{further arguments passed to or from other methods (not currently used).}
}
\value{
An object of class \code{kpca_step}.
}
\description{
For a training set of data, this function uses \code{\link[kernlab]{kpca}} to estimate the loadings for the principal components in the kernel space. This transformation only compute the required statistics for kernel PCA while \code{\link{process}} is used to compute the components on specific data sets.
}
\author{
Max Kuhn
}
\concept{
preprocessing pca projection_methods kernel_methods
}
\keyword{datagen}

