% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nzv.R
\name{learn.nzv_step}
\alias{learn.nzv_step}
\title{Determine Which Variables to Remove Using a Training Set.}
\usage{
\method{learn}{nzv_step}(x, data, ...)
}
\arguments{
\item{x}{a \code{nzv_step} object that contains the list of predictors that should be removed.}

\item{data}{a tibble or data frame that contains the training set.}

\item{...}{further arguments passed to or from other methods (not currently used).}
}
\value{
An object of class \code{nzv_step}.
}
\description{
For a training set of data, this function uses \code{\link[caret]{nearZeroVar}} to determine which, if any, columns in the training set have sparse and unbalanced distributions.
}
\author{
Max Kuhn
}
\concept{
preprocessing variable_filters
}
\keyword{datagen}

