% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pca.R
\name{learn.pca_step}
\alias{learn.pca_step}
\title{Estimate the Principal Component Loadings from a Training Set.}
\usage{
\method{learn}{pca_step}(x, data, ...)
}
\arguments{
\item{x}{a \code{pca_step} object that contains the PCA objects.}

\item{data}{a tibble or data frame that contains the training set. These data will be used to compute the loadings that are used when this step is applied.}

\item{...}{further arguments passed to or from other methods (not currently used).}
}
\value{
An object of class \code{pca_step}.
}
\description{
For a training set of data, this function uses \code{\link[stats]{prcomp}} to estimate the loadings for the principal components. This transformation only compute the required statistics for PCA while \code{\link{process}} is used to compute the components on specific data sets. 
#'
}
\author{
Max Kuhn
}
\concept{
preprocessing pca projection_methods
}
\keyword{datagen}

