% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale.R
\name{learn.scale_step}
\alias{learn.scale_step}
\title{Estimate Standard Deviations from a Training Set for Scaling}
\usage{
\method{learn}{scale_step}(x, data, ...)
}
\arguments{
\item{x}{a \code{scale_step} object that specifies which columns will be scaled}

\item{data}{a tibble or data frame that contains the training set.}

\item{...}{further arguments passed to or from other methods (not currently used).}
}
\value{
An object of class \code{scale_step}.
}
\description{
For a training set of data, this function computes the sample standard deviation across numeric columns that require scaling Note that no data are scaled by this function; see  \code{\link{process.scale_step}}.
}
\author{
Max Kuhn
}
\concept{
preprocessing normalization_methods
}
\keyword{datagen}

