% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatialsign.R
\name{learn.spatialsign_step}
\alias{learn.spatialsign_step}
\title{Setup a Training Set for the Spatial Sign}
\usage{
\method{learn}{spatialsign_step}(x, data, ...)
}
\arguments{
\item{x}{a \code{spatialsign_step} object that contains the PCA objects.}

\item{data}{a tibble or data frame that contains the training set. These data will be used to compute the loadings that are used when this step is applied.}

\item{...}{further arguments passed to or from other methods (not currently used).}
}
\value{
An object of class \code{spatialsign_step}.
}
\description{
The spatial sign is a transformation of the original variables that does not require any parameters to be estimated. As such, this function only returns the original object. 
#'
}
\author{
Max Kuhn
}
\concept{
preprocessing projection_methods
}
\keyword{datagen}

