% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{names0}
\alias{names0}
\title{Sequences of Names with Padded Zeros}
\usage{
names0(num, prefix = "x")
}
\arguments{
\item{num}{A single integer for how many elements are created.}

\item{prefix}{A character string that will start each name. .}
}
\value{
A character string of length \code{num}.
}
\description{
This function creates a series of \code{num} names with a common prefix.
  The names are numbered with leading zeros (e.g.
  \code{prefix01}-\code{prefix10} instead of \code{prefix1}-\code{prefix10}).
}
\concept{
string_functions naming_functions
}
\keyword{datagen}
