% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{okc}
\alias{okc}
\title{OkCupid Data}
\source{
Kim, A. Y., and A. Escobedo-Land. 2015. "OkCupid Data for
Introductory Statistics and Data Science Courses." \emph{Journal of
Statistics Education: An International Journal on the Teaching and
Learning of Statistics}.
}
\value{
\item{okc}{a data frame}
}
\description{
These are a sample of columns of users of OkCupid dating website. The data
are from Kim and Escobedo-Land (2015). Permission to use this data set was
explicitly granted by OkCupid.
}
\examples{
data(okc)
str(okc)
}
\keyword{datasets}
