% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selections.R
\name{parse_terms_formula}
\alias{parse_terms_formula}
\title{Parse a Step Function Formula.}
\usage{
parse_terms_formula(f, info)
}
\arguments{
\item{f}{A formula that can inlcude variable names and certain functions. See \code{\link{selections}} for a list of functions.}

\item{info}{A tibble with columns \code{variable}, \code{type}, \code{role}, and \code{source} that represent the current state of the data. The function \code{\link{summary.recipe}} can be used to get this information from a recipe.}
}
\value{
A character string of column names.
}
\description{
\code{parse_terms_formula} takes the \code{terms} element of a step and translates it to a character vector of column names.
}
\examples{
rec <- recipe( ~ ., data = USArrests)
rec <- step_pca(rec, terms = ~ numerics(), num = 3)
parse_terms_formula(~ predictors(), summary(rec))
rec <- learn(rec, training = USArrests)
parse_terms_formula(~ predictors(), summary(rec))
}
\seealso{
\code{\link{recipe}} \code{\link{summary.recipe}} \code{\link{learn.recipe}}
}
\concept{
preprocessing
}
\keyword{datagen}
