% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprec-misc.R
\name{prepare}
\alias{prepare}
\title{Estimate a preprocessing recipe}
\usage{
prepare(x, ...)
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}

Please use \code{\link[=prep]{prep()}} instead.
}
\keyword{internal}
