% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepper.R
\name{prepper}
\alias{prepper}
\title{Wrapper function for preparing recipes within resampling}
\usage{
prepper(split_obj, recipe, ...)
}
\arguments{
\item{split_obj}{An \code{rplit} object}

\item{recipe}{An untrained \code{recipe} object.}

\item{...}{Arguments to pass to \code{prep} such as \code{verbose} or \code{retain}.}
}
\description{
When working with the \pkg{rsample} package, a simple recipe must be
\emph{prepared} using the \code{prep} function first. When using recipes with
\pkg{rsample} it is helpful to have a function that can prepare a recipe
across a series of \code{split} objects that are produced in this package.
\code{prepper} is a wrapper function around \code{\link[=prep]{prep()}} that can be used to do this.
See the vignette on "Recipes and rsample" for an example.
}
\details{
\code{prepper()} sets the underlying \code{\link[=prep]{prep()}} argument \code{fresh} to \code{TRUE}.
}
