% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recipe.R
\name{print.recipe}
\alias{print.recipe}
\title{Print a Recipe}
\usage{
\method{print}{recipe}(x, form_width = 30, ...)
}
\arguments{
\item{x}{A \code{recipe} object}

\item{form_width}{The number of characters used to print the variables or
terms in a formula}

\item{...}{further arguments passed to or from other methods (not currently
used).}
}
\value{
The original object (invisibly)
}
\description{
Print a Recipe
}
