% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{printer}
\alias{printer}
\title{Printing Workhorse Function}
\usage{
printer(tr_obj = NULL, untr_obj = NULL, trained = FALSE, width = max(20,
  options()$width - 30))
}
\arguments{
\item{tr_obj}{A character vector of names that have been
resolved during preparing the recipe (e.g. the \code{columns} object
of \code{\link[=step_log]{step_log()}}).}

\item{untr_obj}{An oject of selectors prior to prepping the
recipe (e.g. \code{terms} in most steps).}

\item{trained}{A logical for whether the step has been trained.}

\item{width}{An integer denoting where the output should be wrapped.}
}
\value{
`NULL``, invisibly.
}
\description{
This internal function is used for printing steps.
}
\keyword{internal}
