% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bag_imp.R
\name{process.bagimpute_step}
\alias{process.bagimpute_step}
\title{Impute Missing Data a Data Set using Bagged Trees.}
\usage{
\method{process}{bagimpute_step}(x, data, ...)
}
\arguments{
\item{x}{A trained \code{bagimpute_step} object.}

\item{data}{A tibble or data frame that will be imputed.}

\item{...}{further arguments passed to or from other methods (not currently used).}
}
\value{
A tibble of processed data.
}
\description{
For a trained \code{bagimpute_step} object, this function imputes data based on models created on the training set.
}
\author{
Max Kuhn
}
\concept{
preprocessing imputation
}
\keyword{datagen}

