% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/center.R
\name{process.center_step}
\alias{process.center_step}
\title{Center Variables in a Data Set.}
\usage{
\method{process}{center_step}(x, data, ...)
}
\arguments{
\item{x}{A trained \code{center_step} object.}

\item{data}{A tibble or data frame that has numeric variables that will be centered.}

\item{...}{further arguments passed to or from other methods (not currently used).}
}
\value{
A tibble of processed data.
}
\description{
For a trained \code{center_step} object, this function will subtract the sample mean obtained from the training set from column in the current data.
}
\author{
Max Kuhn
}
\concept{
preprocessing normalization_methods
}
\keyword{datagen}

