% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dummy.R
\name{process.dummy_step}
\alias{process.dummy_step}
\title{Generate dummy variables in a data set.}
\usage{
\method{process}{dummy_step}(x, data, ...)
}
\arguments{
\item{x}{A trained \code{step_dummy} object.}

\item{data}{A tibble or data frame that has nominal variables that will be converted to dumy variables.}

\item{...}{further arguments passed to or from other methods (not currently used).}
}
\value{
A tibble of processed data.
}
\description{
For a trained \code{step_dummy} object, this function can be used to apply the process of creating dummy variables to any data set. This creates new columns in the data set and removes the original column(s).
}
\author{
Max Kuhn
}
\concept{
preprocessing dummy_variables model_specification
}
\keyword{datagen}

