% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interactions.R
\name{process.interact_step}
\alias{process.interact_step}
\title{Create Interactions in a Data Set.}
\usage{
\method{process}{interact_step}(x, data, ...)
}
\arguments{
\item{x}{A trained \code{interact_step} object.}

\item{data}{A tibble or data frame that has numeric variables for the interactions.}

\item{...}{further arguments passed to or from other methods (not currently used).}
}
\value{
A tibble of processed data.
}
\description{
For a trained \code{interact_step} object, this function will augment the current data with columns containing the interactions.
}
\author{
Max Kuhn
}
\concept{
preprocessing
}
\keyword{datagen}

