% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kpca.R
\name{process.kpca_step}
\alias{process.kpca_step}
\title{Compute the Kernel Principal Components for a Data Set.}
\usage{
\method{process}{kpca_step}(x, data, ...)
}
\arguments{
\item{x}{A trained \code{kpca_step} object.}

\item{data}{A tibble or data frame that has numeric variables that will be converted to principal components.}

\item{...}{further arguments passed to or from other methods (not currently used).}
}
\value{
A tibble of processed data.
}
\description{
For a trained \code{kpca_step} object, this function projects the current data into the kernel principal components defined by the training set. This creates new columns in the data set and removes the original columns.
}
\author{
Max Kuhn
}
\concept{
preprocessing pca projection_methods kernel_methods
}
\keyword{datagen}

