% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nzv.R
\name{process.nzv_step}
\alias{process.nzv_step}
\title{Filter Near-Zero Variance Columns form a Data Set.}
\usage{
\method{process}{nzv_step}(x, data, ...)
}
\arguments{
\item{x}{A trained \code{nzv_step} object.}

\item{data}{A tibble or data frame.}

\item{...}{further arguments passed to or from other methods (not currently used).}
}
\value{
A tibble of processed data.
}
\description{
For a trained \code{nzv_step} object, this function potentially removes columns from the data.
}
\author{
Max Kuhn
}
\concept{
preprocessing variable_filters
}
\keyword{datagen}

