% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale.R
\name{process.scale_step}
\alias{process.scale_step}
\title{Scale Variables in a Data Set.}
\usage{
\method{process}{scale_step}(x, data, ...)
}
\arguments{
\item{x}{A trained \code{scale_step} object.}

\item{data}{A tibble or data frame that has numeric variables that will be scaled}

\item{...}{further arguments passed to or from other methods (not currently used).}
}
\value{
A tibble of processed data.
}
\description{
For a trained \code{scale_step} object, this function will divide the sample standard deviation obtained from the training set into the columns in the current data.
}
\author{
Max Kuhn
}
\concept{
preprocessing normalization_methods
}
\keyword{datagen}

