% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatialsign.R
\name{process.spatialsign_step}
\alias{process.spatialsign_step}
\title{Compute the Spatial Sign for a Data Set.}
\usage{
\method{process}{spatialsign_step}(x, data, ...)
}
\arguments{
\item{x}{A trained \code{spatialsign_step} object.}

\item{data}{A tibble or data frame that has numeric variables that will be transformed.}

\item{...}{further arguments passed to or from other methods (not currently used).}
}
\value{
A tibble of processed data.
}
\description{
For a trained \code{spatialsign_step} object, this function projects the current data into a unit sphere. This creates overwrites the original columnms
}
\author{
Max Kuhn
}
\concept{
preprocessing projection_methods
}
\keyword{datagen}

