% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/YeoJohnson.R
\name{yj_trans}
\alias{yj_trans}
\alias{estimate_yj}
\title{Internal Functions}
\usage{
yj_trans(x, lambda, eps = 0.001)

estimate_yj(dat, limits = c(-5, 5), nunique = 5)
}
\description{
These are not to be used directly by the users.
}
\keyword{internal}
