% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{recipes-role-indicator}
\alias{recipes-role-indicator}
\alias{recipes_names_predictors}
\alias{recipes_names_outcomes}
\title{Role indicators}
\usage{
recipes_names_predictors(info)

recipes_names_outcomes(info)
}
\arguments{
\item{info}{data.frame with variable information of columns.}
}
\value{
Character vector of column names.
}
\description{
This helper function is meant to be used in \code{prep()} methods to identify
predictors and outcomes by names.
}
\seealso{
\link{developer_functions}
}
\keyword{internal}
