% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg_check.R
\name{recipes_pkg_check}
\alias{recipes_pkg_check}
\title{Update packages}
\usage{
recipes_pkg_check(pkg = NULL, ...)
}
\arguments{
\item{pkg}{A character string for the package being checked}

\item{...}{Extra arguments to pass to \code{\link[utils:install.packages]{utils::install.packages()}}}
}
\value{
Nothing is returned but a message is printed to the console about
which packages (if any) should be installed along with code to do so.
}
\description{
This will check to see if all required packages are installed.
}
\examples{
\dontrun{
recipes_pkg_check("recipes")
}
}
\seealso{
\link{developer_functions}
}
\keyword{internal}
