% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{recipes_remove_cols}
\alias{recipes_remove_cols}
\title{Removes columns if options apply}
\usage{
recipes_remove_cols(new_data, object, col_names = character())
}
\arguments{
\item{new_data}{A tibble.}

\item{object}{A step object.}

\item{col_names}{A character vector, denoting columns to remove. Will
overwrite \code{object$removals} if set.}
}
\value{
\code{new_data} with column names removed if specified by \code{col_names} or
\code{object$removals}.
}
\description{
This helper function removes columns based on character vectors.
}
\seealso{
\link{developer_functions}
}
\keyword{internal}
