% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{recipes_translate_types}
\alias{recipes_translate_types}
\title{Translates into accepted recipes data types}
\usage{
recipes_translate_types(x)
}
\arguments{
\item{x}{Table object used to determine data types}
}
\description{
Translates into accepted recipes data types
}
\details{
This function should be used for integration with other types of
data sources. It is expected that this method will be used by other packages
to provide a translation of the data types.
}
