% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reexports.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{tidy}
\alias{required_pkgs}
\alias{tunable}
\alias{tune_args}
\alias{extract_parameter_set_dials}
\alias{extract_parameter_dials}
\alias{importance_weights}
\alias{frequency_weights}
\alias{\%>\%}
\alias{extract_fit_time}
\alias{update}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{generics}{\code{\link[generics:required_pkgs]{required_pkgs()}}, \code{\link[generics:tidy]{tidy()}}, \code{\link[generics:tunable]{tunable()}}, \code{\link[generics:tune_args]{tune_args()}}}

  \item{hardhat}{\code{\link[hardhat:extract_fit_time]{extract_fit_time()}}, \code{\link[hardhat:extract_parameter_dials]{extract_parameter_dials()}}, \code{\link[hardhat:extract_parameter_set_dials]{extract_parameter_set_dials()}}, \code{\link[hardhat:frequency_weights]{frequency_weights()}}, \code{\link[hardhat:importance_weights]{importance_weights()}}}

  \item{magrittr}{\code{\link[magrittr:\%>\%]{\%>\%}}}

  \item{stats}{\code{\link[stats:update]{update()}}}
}}

