% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{remove_original_cols}
\alias{remove_original_cols}
\title{Removes original columns if options apply}
\usage{
remove_original_cols(new_data, object, col_names)
}
\arguments{
\item{new_data}{A tibble.}

\item{object}{A step object.}

\item{col_names}{A character vector, denoting columns to remove.}
}
\value{
new_data with \code{col_names} removed if \code{get_keep_original_cols(object) == TRUE} or \code{object$preserve == TRUE}.
}
\description{
This helper function should be used whenever the argument
\code{keep_original_cols} is used in a function.
}
\seealso{
\link{developer_functions}
}
\keyword{internal}
