% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dtplyr_functions.R
\name{schema}
\alias{schema}
\title{Schema (column classes) of the table}
\usage{
schema(x, ...)
}
\arguments{
\item{x}{table}
}
\value{
named character vector of classes for each column
}
\description{
Schema (column classes) of the table
}
