% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy.R
\name{sel2char}
\alias{sel2char}
\title{Convert Selectors to Character}
\usage{
sel2char(x)
}
\arguments{
\item{x}{A list of selectors}
}
\value{
A character vector
}
\description{
This internal function takes a list of selectors (e.g. \code{terms}
in most steps) and returns a character vector version for
printing.
}
\keyword{internal}
