% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selections.R
\name{select_terms}
\alias{select_terms}
\title{Select Terms.}
\usage{
select_terms(info, ...)
}
\arguments{
\item{info}{A tibble with columns \code{variable}, \code{type}, \code{role}, and \code{source} that represent the current state of the data. The function \code{\link{summary.recipe}} can be used to get this information from a recipe.}

\item{...}{One or more selection functions. See \code{\link{selections}} for more details.}
}
\value{
A character string of column names or an error of there are no selectors or if no variables are selected.
}
\description{
This function processes the step function selectors and might be useful when creating custom steps.
}
\seealso{
\code{\link{recipe}} \code{\link{summary.recipe}} \code{\link{learn.recipe}}
}
\concept{
preprocessing
}
\keyword{datagen}
