% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Smithsonian}
\alias{Smithsonian}
\title{Smithsonian Museums}
\source{
https://en.wikipedia.org/wiki/List_of_Smithsonian_museums
}
\value{
\item{Smithsonian}{a tibble}
}
\description{
Geocodes for the Smithsonian museums (circa 2018).
}
\examples{
data(Smithsonian)
Smithsonian
}
\keyword{datasets}
