% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alr.R
\name{step_alr}
\alias{step_alr}
\alias{tidy.step_alr}
\title{Additive Log Ratio Transformation of Compositional Data}
\usage{
step_alr(recipe, ..., role = NA, trained = FALSE, columns = NULL,
  ivar = NULL, prefix = "ilr_", skip = FALSE)

\method{tidy}{step_alr}(x, ...)
}
\arguments{
\item{recipe}{A recipe object. The step will be added to the
sequence of operations for this recipe.}

\item{...}{One or more selector functions to choose which
variables are affected by the step. See \code{\link[=selections]{selections()}}
for more details. For the \code{tidy} method, these are not
currently used.}

\item{role}{Not used by this step since no new variables are
created.}

\item{trained}{A logical to indicate if the quantities for
preprocessing have been estimated.}

\item{skip}{A logical. Should the step be skipped when the
recipe is baked by \code{\link[=bake.recipe]{bake.recipe()}}? While all operations are baked
when \code{\link[=prep.recipe]{prep.recipe()}} is run, some operations may not be able to be
conducted on new data (e.g. processing the outcome variable(s)).
Care should be taken when using \code{skip = TRUE} as it may affect
the computations for subsequent operations}

\item{x}{A \code{step_alr} object.}

\item{means}{A named numeric vector of means. This is
\code{NULL} until computed by \code{\link[=prep.recipe]{prep.recipe()}}.}

\item{na.rm}{A logical value indicating whether \code{NA}
values should be removed during computations.}
}
\value{
An updated version of \code{recipe} with the new step
added to the sequence of existing steps (if any). For the
\code{tidy} method, a tibble with columns \code{terms} (the
selectors or variables selected) and \code{value} (the means).
}
\description{
\code{step_alr} creates a \emph{specification} of a recipe
step that will transform a dataset of compositions using the
additive log ratio (ALR) transform.
}
\details{
The ALR transform maps a composition of a $D$ simplex into
a $D - 1$ Euclidian vector. The alr is given by,
\deqn{\mathrm{alr}(x)_i = \log \left(\frac{x_i}{x_D} \right)}{alr(x)_i = log(x_i / x_D)} .

ilring data means that the average of a variable is
subtracted from the data. \code{step_alr} estimates the
variable means from the data used in the \code{training}
argument of \code{prep.recipe}. \code{bake.recipe} then applies
the ilring to new data sets using these means.
}
\examples{
data(biomass)

library('compositions')
data(Hydrochem)

recipe(~ ., data = Hydrochem) \%>\%
  step_alr(all_numeric(), -Code, -Site) \%>\%
  prep() \%>\%
  bake(newdata = Hydrochem)
}
\seealso{
\code{\link[=recipe]{recipe()}} \code{\link[=prep.recipe]{prep.recipe()}}
\code{\link[=bake.recipe]{bake.recipe()}}
}
\concept{
preprocessing normalization_methods
}
\keyword{datagen}
