% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BoxCox.R
\name{step_BoxCox}
\alias{learn.step_BoxCox}
\alias{process.step_BoxCox}
\alias{step_BoxCox}
\title{Box-Cox Transformation for Non-Negative Data}
\usage{
step_BoxCox(recipe, terms, role = NA, trained = FALSE, lambdas = NULL,
  limits = c(-5, 5), nunique = 5)

\method{learn}{step_BoxCox}(x, training, ...)

\method{process}{step_BoxCox}(object, newdata, ...)
}
\arguments{
\item{recipe}{A recipe object. The step will be added to the sequence of operations for this recipe.}

\item{terms}{A representation of the variables or terms that will be transformed.}

\item{role}{Not used by this step since no new variables are created.}

\item{trained}{A logical to indicate if the quantities for preprocessing have been estimated.}

\item{lambdas}{A numeric vector of transformation values. This is \code{NULL} until computed by \code{\link{learn.step_BoxCox}}.}

\item{limits}{A length 2 numeric vector defining the range to compute the transformation parameter lambda.}

\item{nunique}{An integer where data that have less possible values will not be evaluate for a transformation}

\item{x}{a \code{step_BoxCox} object that specifies which columns will be transformed}

\item{training}{a tibble or data frame that contains the training set.}

\item{...}{further arguments passed to or from other methods (not currently used).}

\item{object}{A trained step object.}

\item{newdata}{A tibble or data frame that has numeric variables that will be transformed}
}
\value{
\code{step_BoxCox} and \code{learn.step_BoxCox} return objects of class \code{step_BoxCox}.

\code{process.step_BoxCox} returns a tibble of processed data.
}
\description{
\code{step_BoxCox} creates a \emph{specification} of a recipe step that will transform data using a simple Box-Cox transformation.

For a training set of data, \code{learn.step_BoxCox} estimates the simple Box-Cox transformation.

\code{process.step_BoxCox} is used to transform columns on specific data sets. This replaces values in the original columns.
}
\concept{
preprocessing transformation_methods
}
\keyword{datagen}

