% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/center.R
\name{step_center}
\alias{step_center}
\title{Centering Numeric Data}
\usage{
step_center(recipe, terms, role = NA, trained = FALSE, means = NULL)
}
\arguments{
\item{recipe}{A recipe object. The step will be added to the sequence of operations for this recipe.}

\item{terms}{A formula that represents the variables or terms that will be processed. The raw variable names can be used or \pkg{dplyr} selection tools. See \code{\link{selections}} for more details.}

\item{role}{Not used by this step since no new variables are created.}

\item{trained}{A logical to indicate if the quantities for preprocessing have been estimated.}

\item{means}{A named numeric vector of means. This is \code{NULL} until computed by \code{\link{learn.recipe}}.}
}
\value{
\code{step_center} returns an object of class \code{step_center}.
}
\description{
\code{step_center} creates a \emph{specification} of a recipe step that will normalize numeric data to have a mean of zero.
}
\concept{
preprocessing normalization_methods
}
\keyword{datagen}

