% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downsample.R
\name{step_downsample}
\alias{step_downsample}
\title{Down-Sample a Data Set Based on a Factor Variable}
\usage{
step_downsample(...)
}
\arguments{
\item{...}{Arguments.}
}
\description{
This function is not located in the \pkg{themis} package.
}
\keyword{datagen}
