% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/harmonic.R
\name{step_harmonic}
\alias{step_harmonic}
\title{Add sin and cos terms for harmonic analysis}
\usage{
step_harmonic(
  recipe,
  ...,
  role = NA,
  trained = FALSE,
  frequency = NA_real_,
  period = NA_real_,
  cycle_unit = "day",
  starting_val = NA_real_,
  objects = NULL,
  skip = FALSE,
  id = rand_id("harmonic")
)
}
\arguments{
\item{recipe}{A recipe object. The step will be added to the sequence of
operations for this recipe.}

\item{...}{One or more selector functions to choose which variables are
affected by the step. See \code{\link[=selections]{selections()}} for more details.}

\item{role}{For model terms created by this step, what analysis
role should they be assigned?. By default, the function assumes
that the new columns created from the original variables will be
used as predictors in a model.}

\item{trained}{A logical to indicate if the quantities for preprocessing
have been estimated. Again included for consistency.}

\item{frequency}{A numeric vector with at least one value if using frequency
to specify the cycle size.  The value must be greater than zero and finite.}

\item{period}{A numeric vector with at least one value if using period to
specify the cycle size. The value must be greater than zero and finite.}

\item{cycle_unit}{Character string to indicate the units for a cycle.
Possible values are:
'year' (31556926 seconds), 'month_synodic' (2360592 seconds),
'month_sidereal' (2551443 seconds), 'month_average' (2629744 seconds),
'week', day', 'hour', 'minute', 'second', 'sample'.}

\item{starting_val}{either numeric, Date or POSIXt value that indicates the
reference point for the sin and cos curves.  This parameter may be
specified to increase control over the signal phase.  If starting_val is
not specified the default is the start of the data series, or the start of
the cycle (i.e. the beginning of the hour, day, week).}

\item{objects}{Statistics are stored here once this step has
been trained by \code{\link[=prep.recipe]{prep.recipe()}}.}
}
\value{
An updated version of \code{recipe} with the
new step added to the sequence of existing steps (if any).
}
\description{
\code{step_harmonic} creates a \emph{specification} of a recipe step that
will add sin and cos terms for harmonic analysis.
}
\examples{
library(ggplot2, quietly = TRUE)

data(sunspot.year)
sunspots <- tibble(yr = 1700:1988,
                   n_sunspot = sunspot.year,
                   type = "measured")[1:75,]

# sunspots period is around 11 years, sample spacing is one year
rec <- recipe(n_sunspot ~ yr,
              data = sunspots) \%>\%
              step_harmonic(yr, period = 11, cycle_unit = "sample") \%>\%
              prep() \%>\%
              bake(new_data = NULL)

 fit <- lm(n_sunspot~yr_sin_p_11+yr_cos_p_11, rec)

 preds <- tibble(yr = sunspots$yr,
                 n_sunspot = fit$fitted.values,
                 type = 'predicted')

 ggplot(rbind(sunspots, preds), aes(x = yr, y = n_sunspot, color = type)) +
   geom_line()

}
\seealso{
\code{\link[=recipe]{recipe()}} \code{\link[=prep.recipe]{prep.recipe()}} \code{\link[=bake.recipe]{bake.recipe()}}
}
