% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hyperbolic.R
\name{step_hyperbolic}
\alias{step_hyperbolic}
\title{Hyperbolic Transformations}
\usage{
step_hyperbolic(recipe, terms, role = NA, trained = FALSE, func = "sin",
  inverse = TRUE, vars = NULL)
}
\arguments{
\item{recipe}{A recipe object. The step will be added to the sequence of operations for this recipe.}

\item{terms}{A formula that represents the variables or terms that will be processed. The raw variable names can be used or \pkg{dplyr} selection tools. See \code{\link{selections}} for more details.}

\item{role}{Not used by this step since no new variables are created.}

\item{trained}{A logical to indicate if the quantities for preprocessing have been estimated.}

\item{func}{A character value for the function. Valid values are "sin", "cos", or "tan".}

\item{inverse}{A logical: should the inverse function be used?}

\item{vars}{A character string of variable names that will be (eventually) populated by the \code{terms} argument.}
}
\value{
\code{step_hyperbolic} returns an object of class \code{step_hyperbolic}.
}
\description{
\code{step_hyperbolic} creates a \emph{specification} of a recipe step that will transform data using a hyperbolic function.
}
\concept{
preprocessing transformation_methods
}
\keyword{datagen}

