% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ica.R
\name{step_ica}
\alias{step_ica}
\title{ICA Signal Extraction.}
\usage{
step_ica(recipe, terms, role = "predictor", trained = FALSE, num = 5,
  options = list(), res = NULL)
}
\arguments{
\item{recipe}{A recipe object. The step will be added to the sequence of operations for this recipe.}

\item{terms}{A representation of the variables or terms that will be used to compute the components.}

\item{role}{For model terms created by this step, what analysis role should they be assigned?. By default, the function assumes that the new independent component columns created by the original variables will be used as predictors in a model.}

\item{trained}{A logical to indicate if the quantities for preprocessing have been estimated.}

\item{num}{The number of ICA components to retain as new predictors. If \code{num} is greater than the number of columns or the number of possible components, a smaller value will be used.}

\item{options}{A list of options to \code{\link[fastICA]{fastICA}}. No defaults are set here. \bold{Note} that the arguments \code{X} and \code{n.comp} should not be passed here.}

\item{res}{The \code{\link[fastICA]{fastICA}} object is stored here once this preprocessing step has be trained by \code{\link{learn.recipe}}.}
}
\value{
\code{step_ica} returns an object of class \code{step_ica}.
}
\description{
\code{step_ica} creates a \emph{specification} of a recipe step that will convert numeric data into one or more independent components.
}
\concept{
preprocessing ica projection_methods
}
\keyword{datagen}

