% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ilr.R
\name{step_ilr}
\alias{step_ilr}
\alias{tidy.step_ilr}
\title{Isometric Log Ratio Transformation of Compositional Data}
\usage{
step_ilr(recipe, ..., role = NA, trained = FALSE,
  V = compositions::ilrBase, prefix = "ilr_", skip = FALSE)

\method{tidy}{step_ilr}(x, ...)
}
\arguments{
\item{recipe}{A recipe object. The step will be added to the
sequence of operations for this recipe.}

\item{...}{One or more selector functions to choose which
variables are affected by the step. See \code{\link[=selections]{selections()}}
for more details. For the \code{tidy} method, these are not
currently used.}

\item{role}{Not used by this step since no new variables are
created.}

\item{trained}{A logical to indicate if the quantities for
preprocessing have been estimated.}

\item{skip}{A logical. Should the step be skipped when the
recipe is baked by \code{\link[=bake.recipe]{bake.recipe()}}? While all operations are baked
when \code{\link[=prep.recipe]{prep.recipe()}} is run, some operations may not be able to be
conducted on new data (e.g. processing the outcome variable(s)).
Care should be taken when using \code{skip = TRUE} as it may affect
the computations for subsequent operations}

\item{x}{A \code{step_ilr} object.}

\item{means}{A named numeric vector of means. This is
\code{NULL} until computed by \code{\link[=prep.recipe]{prep.recipe()}}.}

\item{na.rm}{A logical value indicating whether \code{NA}
values should be removed during computations.}
}
\value{
An updated version of \code{recipe} with the new step
added to the sequence of existing steps (if any). For the
\code{tidy} method, a tibble with columns \code{terms} (the
selectors or variables selected) and \code{value} (the means).
}
\description{
\code{step_ilr} creates a \emph{specification} of a recipe
step that will apply the isometric log-ratio transformation,
(ILR) compositional data to have a mean of zero.
}
\details{
ilring data means that the average of a variable is
subtracted from the data. \code{step_ilr} estimates the
variable means from the data used in the \code{training}
argument of \code{prep.recipe}. \code{bake.recipe} then applies
the ilring to new data sets using these means.
}
\examples{
data(biomass)

library('compositions')
data(Hydrochem)

recipe(~ ., data = Hydrochem) \%>\%
  step_ilr(all_numeric(), -Code, -Site) \%>\%
  prep() \%>\%
  bake(newdata = Hydrochem)
}
\seealso{
\code{\link[=recipe]{recipe()}} \code{\link[=prep.recipe]{prep.recipe()}}
\code{\link[=bake.recipe]{bake.recipe()}}
}
\concept{
preprocessing normalization_methods
}
\keyword{datagen}
