% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intercept.R
\name{step_intercept}
\alias{step_intercept}
\title{Add intercept}
\usage{
step_intercept(recipe, ..., role = "predictor", trained = FALSE)
}
\arguments{
\item{recipe}{A recipe object. The step will be added to the sequence of 
operations for this recipe.}

\item{...}{Name of the intercept column. By default it is \code{Intercept}.}

\item{role}{By default predictor.}

\item{trained}{A logical to indicate if the quantities for preprocessing 
have been estimated.}
}
\value{
An updated version of \code{recipe} with the
  new step added to the sequence of existing steps (if any).
}
\description{
\code{step_intercept} adds an intercept if there isn't already one.
}
\examples{
rec <- recipe(Sepal.Width ~ Sepal.Length + Petal.Length, data = iris)

intercept_trans <- rec  \%>\%
  step_intercept()

intercept_obj <- prepare(intercept_trans, retain = TRUE)

head(juice(intercept_obj))
}
\concept{
preprocessing transformation_methods
}
\keyword{datagen}
