% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/invlogit.R
\name{step_invlogit}
\alias{step_invlogit}
\title{Inverse Logit Transformation}
\usage{
step_invlogit(recipe, terms, role = NA, trained = FALSE, vars = NULL)
}
\arguments{
\item{recipe}{A recipe object. The step will be added to the sequence of operations for this recipe.}

\item{terms}{A formula that represents the variables or terms that will be processed. The raw variable names can be used or \pkg{dplyr} selection tools. See \code{\link{selections}} for more details.}

\item{role}{Not used by this step since no new variables are created.}

\item{trained}{A logical to indicate if the quantities for preprocessing have been estimated.}

\item{vars}{A character string of variable names that will be (eventually) populated by the \code{terms} argument.}
}
\value{
\code{step_invlogit} returns an object of class \code{step_invlogit}.
}
\description{
\code{step_invlogit} creates a \emph{specification} of a recipe step that will transform the data from real values to be between zero and one.
}
\concept{
preprocessing transformation_methods
}
\keyword{datagen}

