% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isomap.R
\name{step_isomap}
\alias{learn.step_isomap}
\alias{process.step_isomap}
\alias{step_isomap}
\title{Isomap Embedding.}
\usage{
step_isomap(recipe, terms, role = "predictor", trained = FALSE, num = 5,
  options = list(knn = 50, .mute = c("message", "output")), res = NULL)

\method{learn}{step_isomap}(x, training, ...)

\method{process}{step_isomap}(object, newdata, ...)
}
\arguments{
\item{recipe}{A recipe object. The step will be added to the sequence of operations for this recipe.}

\item{terms}{A representation of the variables or terms that will be used to compute the dimensions.}

\item{role}{For model terms created by this step, what analysis role should they be assigned?. By default, the function assumes that the new dimension columns created by the original variables will be used as predictors in a model.}

\item{trained}{A logical to indicate if the quantities for preprocessing have been estimated.}

\item{num}{The number of isomap dimensions to retain as new predictors. If \code{num} is greater than the number of columns or the number of possible dimensions, a smaller value will be used.}

\item{options}{A list of options to \code{\link[dimRed]{Isomap}}.}

\item{res}{The \code{\link[dimRed]{Isomap}} object is stored here once this preprocessing step has be trained by \code{\link{learn.step_isomap}}.}

\item{x}{A \code{step_isomap} object that contains the Isomap specifications.}

\item{training}{a tibble or data frame that contains the training set. These data will be used to compute the mappings that are used when this step is applied.}

\item{...}{further arguments passed to or from other methods (not currently used).}

\item{object}{A trained step object.}

\item{newdata}{A tibble or data frame that has numeric variables that will be converted to new dimensions.}
}
\value{
\code{step_isomap} and \code{learn.step_isomap} return objects of class \code{step_isomap}.

\code{process.step_isomap} returns a tibble of processed data.
}
\description{
\code{step_isomap} creates a \emph{specification} of a recipe step that will convert numeric data into one or more new dimensions.

For a training set of data, \code{learn.step_isomap} estimates the mappings for the new dimensions. This transformation only computes the required statistics for the Isomap approximation.

\code{process.step_isomap} is used to compute the new dimensions on specific data sets. This creates new columns in the data set and removes the original columns.
}
\concept{
preprocessing isomap projection_methods
}
\keyword{datagen}

