% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/naindicate.R
\name{step_na_indicate}
\alias{step_na_indicate}
\alias{tidy.step_shadow_missing}
\title{Create Missing Data Column Indicators}
\usage{
step_na_indicate(
  recipe,
  ...,
  role = "predictor",
  trained = FALSE,
  columns = NULL,
  prefix = NULL,
  skip = FALSE,
  id = rand_id("na_indicate")
)

\method{tidy}{step_shadow_missing}(x, ...)
}
\arguments{
\item{...}{One or more selector functions to choose which variables are
affected by the step. See \code{\link[=selections]{selections()}} for more details. For the \code{tidy}
method, these are not currently used.}

\item{role}{For model terms created by this step, what analysis
role should they be assigned?. By default, the function assumes
that the new na indicator columns created from the original
variables will be used as predictors in a model.}

\item{trained}{A logical for whether the selectors in \code{...}
have been resolved by \code{\link[=prep]{prep()}}.}

\item{columns}{A character string of variable names that will
be populated (eventually) by the terms argument.}

\item{prefix}{A character string that will be the prefix to the
resulting new variables. See notes below.}

\item{x}{A \code{step_na_indicate} object.}
}
\value{
An updated version of \code{recipe} with the new step added to the
sequence of existing steps (if any). For the \code{tidy} method, a tibble with
columns \code{terms} (the selectors or variables selected) and \code{model} (the
median value).
}
\description{
\code{step_na_indicate} creates a \emph{specification} of a recipe step that will
create and append additional binary columns to the dataset, to indicate
which observations are missing.
}
\examples{
library(modeldata)
data("credit_data")

## missing data per column
vapply(credit_data, function(x) mean(is.na(x)), c(num = 0))

set.seed(342)
in_training <- sample(1:nrow(credit_data), 2000)

credit_tr <- credit_data[ in_training, ]
credit_te <- credit_data[-in_training, ]

rec <- recipe(Price ~ ., data = credit_tr)

impute_rec <- rec \%>\%
  step_na_indicate(Income, Assets, Debt)

imp_models <- prep(impute_rec, training = credit_tr)

imputed_te <- bake(imp_models, new_data = credit_te, everything())
}
\concept{imputation}
\concept{preprocessing}
\keyword{datagen}
