% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/naomit.R
\name{step_naomit}
\alias{step_naomit}
\alias{tidy.step_naomit}
\title{Remove observations with missing values}
\usage{
step_naomit(recipe, ..., role = NA, trained = FALSE)

\method{tidy}{step_naomit}(x, ...)
}
\arguments{
\item{recipe}{A recipe object. The step will be added to the sequence of
operations for this recipe.}

\item{...}{One or more selector functions to choose which
variables will be used to create the dummy variables. See
\code{\link[=selections]{selections()}} for more details. The selected
variables must be factors.}

\item{role}{Unused, include for consistency with other steps.}

\item{trained}{A logical to indicate if the quantities for preprocessing
have been estimated. Again included for consistency.}

\item{x}{A \code{step_naomit} object.}
}
\value{
An updated version of \code{recipe} with the
new step added to the sequence of existing steps (if any).
}
\description{
\code{step_naomit} creates a \emph{specification} of a recipe step that
will add remove observations (rows of data) if they contain NA
or NaN values.
}
\examples{

TODO

}
\seealso{
\code{\link[=recipe]{recipe()}} \code{\link[=prep.recipe]{prep.recipe()}} \code{\link[=bake.recipe]{bake.recipe()}}
}
