% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nzv.R
\name{step_nzv}
\alias{step_nzv}
\title{Near-Zero Variance Filter.}
\usage{
step_nzv(recipe, terms, role = NA, trained = FALSE, options = list(freqCut
  = 95/5, uniqueCut = 10), removals = NULL)
}
\arguments{
\item{recipe}{A recipe object. The step will be added to the sequence of operations for this recipe.}

\item{terms}{A representation of the variables or terms that will evaluated by the filtering process.}

\item{role}{Not used by this step since no new variables are created.}

\item{trained}{A logical to indicate if the quantities for preprocessing have been estimated.}

\item{options}{A list of options for \code{\link[caret]{nearZeroVar}}. \bold{Note} that the arguments \code{data} and \code{names} should not be included in this list.}

\item{removals}{A character string that contains the names of columns that should be removed. These values are not determined until \code{\link{learn.recipe}} is called.}
}
\value{
\code{step_nzv}  returns an object of class \code{step_nzv}.
}
\description{
\code{step_nzv} creates a \emph{specification} of a recipe step that will potentially remove variables that are highly sparse and unbalanced.
}
\concept{
preprocessing variable_filters
}
\keyword{datagen}

