% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatialsign.R
\name{step_spatialsign}
\alias{step_spatialsign}
\title{Spatial Sign Preprocessing.}
\usage{
step_spatialsign(recipe, terms, role = "predictor", trained = FALSE,
  vars = NULL)
}
\arguments{
\item{recipe}{A recipe object. The step will be added to the sequence of operations for this recipe.}

\item{terms}{A representation of the variables or terms that will be used for the normalization.}

\item{role}{For model terms created by this step, what analysis role should they be assigned?}

\item{trained}{A logical to indicate if the quantities for preprocessing have been estimated.}

\item{vars}{A character string of variable names that will be (eventually) populated by the \code{terms} argument.}
}
\value{
\code{step_spatialsign} returns an object of class \code{step_spatialsign}.
}
\description{
\code{step_spatialsign} is a \emph{specification} of a recipe step that will convert numeric data into a projection on to a unit sphere.
}
\concept{
preprocessing projection_methods
}
\keyword{datagen}

