% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upsample.R
\name{step_upsample}
\alias{step_upsample}
\title{Up-Sample a Data Set Based on a Factor Variable}
\usage{
step_upsample(...)
}
\arguments{
\item{...}{Arguments.}
}
\description{
This function is not located in the \pkg{themis} package.
}
\keyword{datagen}
