% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xgb_binning.R
\name{step_xgb_binning}
\alias{step_xgb_binning}
\title{Discretize numeric variables with XgBoost}
\usage{
step_xgb_binning(recipe, ..., role = NA, trained = FALSE,
  outcome = NULL, rules = NULL, prefix = "xgb_bin", skip = FALSE,
  id = rand_id("xgb_binning"))
}
\arguments{
\item{recipe}{A recipe object. The step will be added to the sequence of
operations for this recipe.}

\item{...}{One or more selector functions to choose which variables are
affected by the step. See \code{\link[=selections]{selections()}} for more details.}

\item{role}{Defaults to \code{"predictor"}.}

\item{trained}{A logical to indicate if the quantities for preprocessing
have been estimated.}

\item{outcome}{A character string to indicate the outcome variable to train XbBoost models.}

\item{rules}{The splitting rules of the best XgBoost tree to retain for each variable.}

\item{prefix}{A character string that will be the prefix to the resulting new variables.
Defaults to "xgb_bin".}
}
\value{
An updated version of \code{recipe} with the
new step added to the sequence of existing steps (if any).
}
\description{
\code{step_xgb_binning} creates a \emph{specification} of a recipe
step that will discretize numeric data (e.g. integeres or doubles)
into bins in a supervised way using an XgBoost model.
}
\details{
XgBoost is a supervised version of principal component
analysis that requires one or more numeric outcomes to compute
the new features. It is advised to impute missing values before
this step.

This step requires the \pkg{xgboost} package. If not installed, the
step will stop with a note about installing the package.

Note that the original variables will be replaced with the new bins.
The new variables will have names that begin with \code{prefix} followed
by their original name.
}
\examples{
data(credit_data)
library(rsample)

split <- initial_split(credit_data, strata = "Status")

credit_data_tr <- training(split)
credit_data_te <- testing(split)

xgb_rec <- recipe(Status ~ ., data = credit_data_tr) \%>\%
  step_medianimpute(all_numeric()) \%>\%
  step_xgb_binning(all_numeric(), outcome = "Status")

xgb_rec <- prep(xgb_rec, training = credit_data_tr, retain = TRUE)

xgb_test_bins <- bake(xgb_rec, credit_data_te)
}
\seealso{
\code{\link[=recipe]{recipe()}} \code{\link[=prep.recipe]{prep.recipe()}} \code{\link[=bake.recipe]{bake.recipe()}}
}
\concept{discretization}
\concept{factors}
\concept{preprocessing}
\concept{xgboost}
\keyword{binning}
