% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/YeoJohnson.R
\name{step_YeoJohnson}
\alias{step_YeoJohnson}
\title{Yeo-Johnson Transformation}
\usage{
step_YeoJohnson(recipe, terms, role = NA, trained = FALSE, lambdas = NULL,
  limits = c(-5, 5), nunique = 5)
}
\arguments{
\item{recipe}{A recipe object. The step will be added to the sequence of operations for this recipe.}

\item{terms}{A formula that represents the variables or terms that will be processed. The raw variable names can be used or \pkg{dplyr} selection tools. See \code{\link{selections}} for more details.}

\item{role}{Not used by this step since no new variables are created.}

\item{trained}{A logical to indicate if the quantities for preprocessing have been estimated.}

\item{lambdas}{A numeric vector of transformation values. This is \code{NULL} until computed by \code{\link{learn.recipe}}.}

\item{limits}{A length 2 numeric vector defining the range to compute the transformation parameter lambda.}

\item{nunique}{An integer where data that have less possible values will not be evaluate for a transformation}
}
\value{
\code{step_YeoJohnson} return an object of class \code{step_YeoJohnson}.
}
\description{
\code{step_YeoJohnson} creates a \emph{specification} of a recipe step that will transform data using a simple Yeo-Johnson transformation.
}
\concept{
preprocessing transformation_methods
}
\keyword{datagen}

