% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zv.R
\name{step_zv}
\alias{step_zv}
\title{Zero Variance Filter}
\usage{
step_zv(recipe, ..., role = NA, trained = FALSE, removals = NULL)
}
\arguments{
\item{recipe}{A recipe object. The step will be added to the sequence of
operations for this recipe.}

\item{...}{One or more selector functions to choose which
variables that will evaluated by the filtering. See
\code{\link{selections}} for more details.}

\item{role}{Not used by this step since no new variables are
created.}

\item{trained}{A logical to indicate if the quantities for preprocessing
have been estimated.}

\item{removals}{A character string that contains the names of
columns that should be removed. These values are not determined
until \code{\link{prep.recipe}} is called.}
}
\value{
An updated version of \code{recipe} with the
  new step added to the sequence of existing steps (if any).
}
\description{
\code{step_zv} creates a \emph{specification} of a recipe step
 that will remove variables that contain only a single value.
}
\examples{
data(biomass)

biomass$one_value <- 1

biomass_tr <- biomass[biomass$dataset == "Training",]
biomass_te <- biomass[biomass$dataset == "Testing",]

rec <- recipe(HHV ~ carbon + hydrogen + oxygen +
                    nitrogen + sulfur + one_value,
              data = biomass_tr)

zv_filter <- rec \%>\%
  step_zv(all_predictors())

filter_obj <- prep(zv_filter, training = biomass_tr)

filtered_te <- bake(filter_obj, biomass_te)
any(names(filtered_te) == "one_value")
}
\seealso{
\code{\link{step_nzv}} \code{\link{step_corr}}
  \code{\link{recipe}}
  \code{\link{prep.recipe}} \code{\link{bake.recipe}}
}
\concept{
preprocessing variable_filters
}
\keyword{datagen}
