% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bag_imp.R, R/bs.R, R/corr.R, R/discretize.R,
%   R/downsample.R, R/ica.R, R/isomap.R, R/knn_imp.R, R/kpca_poly.R,
%   R/kpca_rbf.R, R/meanimpute.R, R/nnmf.R, R/ns.R, R/nzv.R, R/other.R, R/pca.R,
%   R/pls.R, R/poly.R, R/rollimpute.R, R/tunable.R, R/upsample.R, R/window.R
\name{tunable.step_bagimpute}
\alias{tunable.step_bagimpute}
\alias{tunable.step_bs}
\alias{tunable.step_corr}
\alias{tunable.step_discretize}
\alias{tunable.step_downsample}
\alias{tunable.step_ica}
\alias{tunable.step_isomap}
\alias{tunable.step_knnimpute}
\alias{tunable.step_kpca_poly}
\alias{tunable.step_kpca_rbf}
\alias{tunable.step_meanimpute}
\alias{tunable.step_nnmf}
\alias{tunable.step_ns}
\alias{tunable.step_nzv}
\alias{tunable.step_other}
\alias{tunable.step_pca}
\alias{tunable.step_pls}
\alias{tunable.step_poly}
\alias{tunable.step_rollimpute}
\alias{tunable.step}
\alias{tunable.check}
\alias{tunable.step_upsample}
\alias{tunable.step_window}
\title{tunable methods for recipes}
\usage{
\method{tunable}{step_bagimpute}(x, ...)

\method{tunable}{step_bs}(x, ...)

\method{tunable}{step_corr}(x, ...)

\method{tunable}{step_discretize}(x, ...)

\method{tunable}{step_downsample}(x, ...)

\method{tunable}{step_ica}(x, ...)

\method{tunable}{step_isomap}(x, ...)

\method{tunable}{step_knnimpute}(x, ...)

\method{tunable}{step_kpca_poly}(x, ...)

\method{tunable}{step_kpca_rbf}(x, ...)

\method{tunable}{step_meanimpute}(x, ...)

\method{tunable}{step_nnmf}(x, ...)

\method{tunable}{step_ns}(x, ...)

\method{tunable}{step_nzv}(x, ...)

\method{tunable}{step_other}(x, ...)

\method{tunable}{step_pca}(x, ...)

\method{tunable}{step_pls}(x, ...)

\method{tunable}{step_poly}(x, ...)

\method{tunable}{step_rollimpute}(x, ...)

\method{tunable}{step}(x, ...)

\method{tunable}{check}(x, ...)

\method{tunable}{step_upsample}(x, ...)

\method{tunable}{step_window}(x, ...)
}
\arguments{
\item{x}{A recipe step object}

\item{...}{Not used.}
}
\value{
A tibble object.
}
\description{
These functions define what parameters \emph{can} be tuned for specific steps.
They also define the recommended objects from the \code{dials} package that can
be used to generate new parameter values and other characteristics.
}
\keyword{internal}
