% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/woe.R
\name{woe_dictionary}
\alias{woe_dictionary}
\title{WoE dictionary of a set of predictor variables upon a given binary outcome}
\usage{
woe_dictionary(.data, outcome, ..., odds_offset = 1e-06)
}
\arguments{
\item{.data}{A tbl. The data.frame where the variables come from.}

\item{outcome}{bare name of the outcome variable with exactly 2 distinct values.}

\item{...}{bare names of predictor variables or selectors accepted by \code{dplyr::select()}.}

\item{odds_offset}{Default to 1e-6. Offset value to avoid -Inf/Inf from predictor
category with only one outcome class. Set to 0 to allow Inf/-Inf.}
}
\value{
a tibble with summaries and woe for every given predictor variable stacked up.
}
\description{
Builds the woe dictionary of a set of predictor variables upon a given binary outcome.
Convenient to make a woe version of the given set of predictor variables and also to allow
one to tweak some woe values by hand.
}
\details{
You can pass a custom dictionary to \code{step_woe()}. It must have the exactly
the same structure of the output of \code{\link[=woe_dictionary]{woe_dictionary()}}. One easy way to do this
is by tweaking an output returned from it.
}
\examples{

mtcars \%>\% woe_dictionary(am, cyl, gear:carb)


}
\references{
Kullback, S. (1959). \emph{Information Theory and Statistics.} Wiley, New York.

Hastie, T., Tibshirani, R. and Friedman, J. (1986). \emph{Elements of Statistical Learning}, Second Edition, Springer, 2009.
}
