% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/steps_and_checks.R
\name{add_step}
\alias{add_step}
\alias{add_check}
\title{Add a New Operation to the Current Recipe}
\usage{
add_step(rec, object)

add_check(rec, object)
}
\arguments{
\item{rec}{A \code{\link[=recipe]{recipe()}}.}

\item{object}{A step or check object.}
}
\value{
A updated \code{\link[=recipe]{recipe()}} with the new operation in the last slot.
}
\description{
\code{add_step()} adds a step to the last location in the recipe. \code{add_check()}
does the same for checks.
}
\seealso{
\link{developer_functions}
}
