% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formula.R
\name{formula.recipe}
\alias{formula.recipe}
\title{Create a formula from a prepared recipe}
\usage{
\method{formula}{recipe}(x, ...)
}
\arguments{
\item{x}{A recipe object that has been prepared.}

\item{...}{Note currently used.}
}
\value{
A formula.
}
\description{
In case a model formula is required, the formula method can be used on a
recipe to show what predictors and outcome(s) could be used.
}
\examples{

formula(recipe(Species + Sepal.Length ~ ., data = iris) |> prep())

iris_rec <- recipe(Species ~ ., data = iris) |>
  step_center(all_numeric()) |>
  prep()
formula(iris_rec)
}
