% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selections.R
\name{has_role}
\alias{has_role}
\alias{is_nominal}
\alias{is_numeric}
\alias{is_outcome}
\alias{is_predictor}
\alias{type_is}
\title{Role Selection}
\usage{
has_role(x = "predictor", roles = NULL)

is_predictor(roles = NULL)

is_outcome(roles = NULL)

type_is(x = "numeric", types = NULL)

is_numeric(types = NULL)

is_nominal(types = NULL)
}
\arguments{
\item{x}{A single character string for the query.}

\item{roles}{A character string of roles for the current set of terms.}

\item{types}{A character string of roles for the current set of data types}
}
\value{
An integer vector.
}
\description{
\code{has_role}, \code{is_predictor}, and \code{is_outcome} can be used to select variables in a formula that have certain roles. Similarly,  \code{type_is}, \code{is_numeric}, and \code{is_nominal} are used to select columns based on their data type. See \code{\link{selections}} for more details.
}
\keyword{datagen}

