% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selections.R
\name{has_role}
\alias{has_role}
\alias{predictors}
\alias{outcomes}
\alias{has_type}
\alias{numerics}
\alias{nominals}
\title{Role Selection}
\usage{
has_role(x = "predictor", roles = NULL)

predictors(roles = NULL)

outcomes(roles = NULL)

has_type(x = "numeric", types = NULL)

numerics(types = NULL)

nominals(types = NULL)
}
\arguments{
\item{x}{A single character string for the query.}

\item{roles}{A character string of roles for the current set of terms.}

\item{types}{A character string of roles for the current set of data types}
}
\value{
An integer vector.
}
\description{
\code{has_role}, \code{predictors}, and \code{outcomes} can be used to select variables in a formula that have certain roles. Similarly,  \code{has_type}, \code{numerics}, and \code{nominals} are used to select columns based on their data type. See \code{\link{selections}} for more details.
}
\keyword{datagen}
