% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ignore_step.R
\name{ignore_step}
\alias{ignore_step}
\title{Ignore step(s) in an existing recipe}
\usage{
ignore_step(recipe, ignore)
}
\arguments{
\item{recipe}{A recipe object. The step will be removed from the sequence of
operations for this recipe.}

\item{ignore}{Character name or integer identifying one or more steps to
remove. There are three options for character arguments: 1. step name (e.g.,
"normalize" without the prefix "step_"), 2. step id code (e.g., a custom id
name originally supplied to the \code{id} argument when the step was defined OR
the automatically generated code that is appended to the step name by
default), or 3. a full id or combination of both 1 and 2 separated by an
underscore. This last option describes the "full ids" automatically generated
by \code{step_} functions when nothing is supplied to the \code{id} argument.}
}
\value{
An updated version of \code{recipe} with all steps named \code{ignore} removed.
}
\description{
Remove recipe steps by function/id name or by step number.
}
\details{
This function is useful when different models have different
pre-processing requirements since it enables ad hoc step removal from a
baseline recipe (e.g., a recipe that satisfies the pre-processing
preference for most but not all models).
}
\examples{
rec <- recipe(Species ~ ., data = iris) \%>\%
  step_rm(Petal.Width, id = "rm_UDLut") \%>\% # example of a default id format
  step_rm(starts_with("Sepal"), id = "custom_id")

ignore_step(rec, "rm")              # ignores all 'rm' steps
ignore_step(rec, "custom_id")       # ignores single step with matching id
ignore_step(rec, rec$steps[[1]]$id) # ignores the first step with its full id
ignore_step(rec, 1)                 # ignores the first step by integer index
}
