% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recipe.R
\name{process}
\alias{process}
\alias{process.recipe}
\title{Apply a Trained Data Recipe}
\usage{
process(x, ...)

\method{process}{recipe}(x, newdata = x$template, ...)
}
\arguments{
\item{x}{A trained object such as a \code{\link{recipe}} with at least one preprocessing step.}

\item{...}{further arguments passed to or from other methods (not currently used).}

\item{newdata}{A data frame or tibble for whom the preprocessing will be applied.}
}
\value{
A tibble that may have different columns than the original columns in \code{newdata}.
}
\description{
For a recipe with at least one preprocessing step that has been trained by \code{\link{learn.recipe}}, apply the computations to new data.
}
\author{
Max Kuhn
}
\concept{
preprocessing model_specification
}
\keyword{datagen}

