% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recipe.R
\name{process}
\alias{process}
\alias{process.recipe}
\alias{process.recipe}
\title{Apply a Trained Data Recipe}
\usage{
process(object, ...)

\method{process}{recipe}(object, newdata = object$template, roles = "all",
  ...)
}
\arguments{
\item{object}{A trained object such as a \code{\link{recipe}} with at least one preprocessing step.}

\item{...}{further arguments passed to or from other methods (not currently used).}

\item{newdata}{A data frame or tibble for whom the preprocessing will be applied.}

\item{roles}{A character vector to choose which types of columns to return (e.g. "predictor"). By default all columns are returned.}
}
\value{
A tibble that may have different columns than the original columns in \code{newdata}.
}
\description{
For a recipe with at least one preprocessing step that has been trained by \code{\link{learn.recipe}}, apply the computations to new data.
}
\details{
\code{\link{process}} takes a trained recipe and applies the operations to a data set to create a design matrix.

If the original data used to train the data are to be processed, time can be saved by using the \code{retain = TRUE} option of \code{\link{learn}} to avoid duplicating the same operations.

A tibble is always returned but can be easily converted to a data frame or matrix as needed.
}
\author{
Max Kuhn
}
\concept{
preprocessing model_specification
}
\keyword{datagen}
