% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reexports.R, R/update.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{tidy}
\alias{required_pkgs}
\alias{tunable}
\alias{\%>\%}
\alias{update}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{generics}{\code{\link[generics]{required_pkgs}}, \code{\link[generics]{tidy}}, \code{\link[generics]{tunable}}}

  \item{magrittr}{\code{\link[magrittr:pipe]{\%>\%}}}

  \item{stats}{\code{\link[stats]{update}}}
}}

