% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classdist_shrunken.R, R/depth.R, R/ica.R,
%   R/isomap.R, R/kpca.R, R/kpca_poly.R, R/kpca_rbf.R, R/mutate.R,
%   R/nnmf.R, R/nnmf_sparse.R, R/pls.R, R/poly_bernstein.R, R/recipe.R,
%   R/spline_b.R, R/spline_convex.R, R/spline_monotone.R, R/spline_natural.R,
%   R/spline_nonnegative.R, R/window.R
\name{required_pkgs.step_classdist_shrunken}
\alias{required_pkgs.step_classdist_shrunken}
\alias{required_pkgs.step_depth}
\alias{required_pkgs.step_ica}
\alias{required_pkgs.step_isomap}
\alias{required_pkgs.step_kpca}
\alias{required_pkgs.step_kpca_poly}
\alias{required_pkgs.step_kpca_rbf}
\alias{required_pkgs.step_mutate}
\alias{required_pkgs.step_nnmf}
\alias{required_pkgs.step_nnmf_sparse}
\alias{required_pkgs.step_pls}
\alias{required_pkgs.step_poly_bernstein}
\alias{required_pkgs.recipe}
\alias{required_pkgs.step}
\alias{required_pkgs.check}
\alias{required_pkgs.step_spline_b}
\alias{required_pkgs.step_spline_convex}
\alias{required_pkgs.step_spline_monotone}
\alias{required_pkgs.step_spline_natural}
\alias{required_pkgs.step_spline_nonnegative}
\alias{required_pkgs.step_window}
\title{S3 methods for tracking which additional packages are needed for steps.}
\usage{
\method{required_pkgs}{step_classdist_shrunken}(x, ...)

\method{required_pkgs}{step_depth}(x, ...)

\method{required_pkgs}{step_ica}(x, ...)

\method{required_pkgs}{step_isomap}(x, ...)

\method{required_pkgs}{step_kpca}(x, ...)

\method{required_pkgs}{step_kpca_poly}(x, ...)

\method{required_pkgs}{step_kpca_rbf}(x, ...)

\method{required_pkgs}{step_mutate}(x, ...)

\method{required_pkgs}{step_nnmf}(x, ...)

\method{required_pkgs}{step_nnmf_sparse}(x, ...)

\method{required_pkgs}{step_pls}(x, ...)

\method{required_pkgs}{step_poly_bernstein}(x, ...)

\method{required_pkgs}{recipe}(x, infra = TRUE, ...)

\method{required_pkgs}{step}(x, ...)

\method{required_pkgs}{check}(x, ...)

\method{required_pkgs}{step_spline_b}(x, ...)

\method{required_pkgs}{step_spline_convex}(x, ...)

\method{required_pkgs}{step_spline_monotone}(x, ...)

\method{required_pkgs}{step_spline_natural}(x, ...)

\method{required_pkgs}{step_spline_nonnegative}(x, ...)

\method{required_pkgs}{step_window}(x, ...)
}
\arguments{
\item{x}{A recipe or recipe step}

\item{infra}{Should recipes itself be included in the result?}
}
\value{
A character vector
}
\description{
S3 methods for tracking which additional packages are needed for steps.
}
\keyword{internal}
