% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/steps_and_checks.R
\name{step}
\alias{step}
\alias{check}
\title{\code{step} sets the class of the \code{step} and \code{check} is for checks.}
\usage{
step(subclass, ..., .prefix = "step_")

check(subclass, ..., .prefix = "check_")
}
\arguments{
\item{subclass}{A character string for the resulting class. For example,
if \code{subclass = "blah"} the step object that is returned has class
\code{step_blah} or \code{step_blah} depending on the context.}

\item{...}{All arguments to the operator that should be returned.}

\item{.prefix}{Prefix to the subclass created.}
}
\value{
An updated step or check with the new class.
}
\description{
\code{step} sets the class of the \code{step} and \code{check} is for checks.
}
\concept{preprocessing}
\keyword{internal}
