% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bag_imp.R
\name{step_bagimpute}
\alias{learn.bagimpute_step}
\alias{process.bagimpute_step}
\alias{step_bagimpute}
\title{Imputed via Bagged Trees.}
\usage{
step_bagimpute(recipe, terms, role = NA, trained = FALSE, models = NULL,
  options = list(nbagg = 25, keepX = FALSE), impute_with = NULL,
  seed_val = sample.int(10^4, 1))

\method{learn}{bagimpute_step}(x, data, ...)

\method{process}{bagimpute_step}(x, data, ...)
}
\arguments{
\item{recipe}{A recipe object. The step will be added to the sequence of operations for this recipe.}

\item{terms}{A representation of the variables or terms that will be imputed.}

\item{role}{Not used by this step since no new variables are created.}

\item{trained}{A logical to indicate if the quantities for preprocessing have been estimated.}

\item{models}{The \code{\link[ipred]{ipredbagg}} objects are stored here once this bagged trees have be trained by \code{\link{learn.bagimpute_step}}.}

\item{options}{A list of options to \code{\link[ipred]{ipredbagg}}. Defaults are set for the arguments \code{nbagg} and \code{keepX} but others can be passed in. \bold{Note} that the arguments \code{X} and \code{y} should not be passed here.}

\item{impute_with}{A representation of the variables that will be used as predictors in the imputation model. If a column is included in both \code{terms} and \code{impute_with}, it will be removed from the latter.}

\item{seed_val}{A integer used to create reproducible models. The same seed is used across all imputation models.}

\item{x}{A \code{bagimpute_step} object that contains the imputation specifications.}

\item{data}{tibble or data frame that contains the training set. These data will be used to compute the tree ensembles that are used when this step is applied.}

\item{...}{further arguments passed to or from other methods (not currently used).}

\item{data}{A tibble or data frame that will be imputed.}
}
\value{
\code{step_bagimpute} and \code{learn.bagimpute_step} return objects of class \code{bagimpute_step}.

\code{process.bagimpute_step} returns a tibble of processed data.
}
\description{
\code{step_bagimpute} creates a \emph{specification} of a recipe step that will create bagged tree models to impute missing data.

For a training set of data, \code{learn.bagimpute_step} creates models that will be used to impute missing data (using \code{\link[ipred]{ipredbagg}}).

\code{process.bagimpute_step} is used to perform the imputation on specific data sets. This replaces values in the original columns.
}
\concept{
preprocessing imputation
}
\keyword{datagen}

