% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bag_imp.R
\name{step_bagimpute}
\alias{step_bagimpute}
\title{Declare Which Variables Will be Imputed via Bagged Trees.}
\usage{
step_bagimpute(recipe, terms, role = NA, models = NULL,
  options = list(nbagg = 25, keepX = FALSE), impute_with = NULL)
}
\arguments{
\item{recipe}{A recipe object. The step will be added to the sequence of operations for this recipe.}

\item{terms}{A representation of the variables or terms that will be imputed.}

\item{role}{Not used by this step since no new variables are created.}

\item{models}{The \code{\link[ipred]{ipredbagg}} objects are stored here once this bagged trees have be trained by \code{\link{learn.bagimpute_step}}.}

\item{options}{A list of options to \code{\link[ipred]{ipredbagg}}. Defaults are set for the arguments \code{nbagg} and \code{keepX} but others can be passed in. \bold{Note} that the arguments \code{X} and \code{y} should not be passed here.}

\item{impute_with}{A representation of the variables that will be used as predictors in the imputation model. If a column is included in both \code{terms} and \code{impute_with}, it will be removed from the latter.}
}
\value{
An object of class \code{bagimpute_step}.
}
\description{
This function is a \emph{specification} of a recipe step that will create bagged tree models to impute missing data.
}
\author{
Max Kuhn
}
\concept{
preprocessing imputation
}
\keyword{datagen}

