% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprec-impute_bag.R
\name{step_bagimpute}
\alias{step_bagimpute}
\title{Impute via bagged trees}
\usage{
step_bagimpute(
  recipe,
  ...,
  role = NA,
  trained = FALSE,
  impute_with = imp_vars(all_predictors()),
  trees = 25,
  models = NULL,
  options = list(keepX = FALSE),
  seed_val = sample.int(10^4, 1),
  skip = FALSE,
  id = rand_id("impute_bag")
)
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}

Please use \code{\link[=step_impute_bag]{step_impute_bag()}} instead.
}
\keyword{internal}
