% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/center.R
\name{step_center}
\alias{step_center}
\title{Declare Variables for Centering.}
\usage{
step_center(recipe, terms, role = NA, means = NULL)
}
\arguments{
\item{recipe}{A recipe object. The step will be added to the sequence of operations for this recipe.}

\item{terms}{A representation of the variables or terms that will be centered.}

\item{role}{Not used by this step since no new variables are created.}

\item{means}{A named numeric vector of means. This is \code{NULL} until computed by \code{\link{learn.center_step}}.}
}
\value{
An object of class \code{center_step}.
}
\description{
This function is a \emph{specification} of a recipe step that will normalize numeric data to have a mean of zero.
}
\author{
Max Kuhn
}
\concept{
preprocessing normalization_methods
}
\keyword{datagen}

