% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dummy.R
\name{step_dummy}
\alias{learn.dummy_step}
\alias{process.dummy_step}
\alias{step_dummy}
\title{Declare Which Variables Should Be Converted to Dummy Variables.}
\usage{
step_dummy(recipe, terms, role = "predictor", trained = FALSE,
  contrast = options("contrasts"), naming = function(var, lvl) paste(var,
  make.names(lvl), sep = "_"), levels = NULL)

\method{learn}{dummy_step}(x, data, ...)

\method{process}{dummy_step}(x, data, ...)
}
\arguments{
\item{recipe}{A recipe object. The step will be added to the sequence of operations for this recipe.}

\item{terms}{A representation of the variables or terms that will be used to create the dummy variables.}

\item{role}{For model terms created by this step, what analysis role should they be assigned?. By default, the function assumes that the binary dummy variable columns created by the original variables will be used as predictors in a model.}

\item{trained}{A logical to indicate if the quantities for preprocessing have been estimated.}

\item{contrast}{A specification for which type of contrast should be used to make a set of full rank dummy variables. See \code{\link[stats]{contrasts}} for more details. \bold{not currently hooked up}}

\item{naming}{A function that defines the naming convention for new binary columns. See Details below.}

\item{levels}{A list that contains the information needed to create dummy variables for each variable contained in \code{terms}. This is \code{NULL} until the step is trained by \code{\link{learn.dummy_step}}.}

\item{x}{a \code{dummy_step} object that specifies which columns will be converted to dummy variables.}

\item{data}{a tibble or data frame that contains the training set. These data will be used to define the dummy variables for all future data when this step is applied.}

\item{...}{further arguments passed to or from other methods (not currently used).}

\item{x}{A trained \code{dummy_step} object.}

\item{data}{A tibble or data frame that has nominal variables that will be converted to dumy variables.}

\item{...}{further arguments passed to or from other methods (not currently used).}
}
\value{
An object of class \code{dummy_step}.

An object of class \code{dummy_step}.

A tibble of processed data.
}
\description{
This function is a \emph{specification} of a recipe step that will convert nominal data (e.g. character or factors) into one or more numeric binary model terms for the levels of the original data. For example, if a factor column in the data set has levels of "red", "green", "blue", the dummy variable process will create two additional columns of 0/1 data for two of those three values (and remove the original column).

For a training set of data, this function enumerates the possible values of the variables so that dummy variables can be created when a specific data set is \emph{processed} (see \code{\link{process.dummy_step}}).

For a trained \code{dummy_step} object, this function can be used to apply the process of creating dummy variables to any data set. This creates new columns in the data set and removes the original column(s).
}
\concept{
preprocessing dummy_variables model_specification dummy_variables variable_encodings

preprocessing dummy_variables model_specification dummy_variables variable_encodings

preprocessing dummy_variables model_specification dummy_variables variable_encodings
}
\keyword{datagen}

