% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dummy.R
\name{step_dummy}
\alias{learn.step_dummy}
\alias{process.step_dummy}
\alias{step_dummy}
\title{Dummy Variables Creation.}
\usage{
step_dummy(recipe, terms, role = "predictor", trained = FALSE,
  contrast = options("contrasts"), naming = function(var, lvl) paste(var,
  make.names(lvl), sep = "_"), levels = NULL)

\method{learn}{step_dummy}(x, training, info = NULL, ...)

\method{process}{step_dummy}(object, newdata, ...)
}
\arguments{
\item{recipe}{A recipe object. The step will be added to the sequence of operations for this recipe.}

\item{terms}{A representation of the variables or terms that will be used to create the dummy variables.}

\item{role}{For model terms created by this step, what analysis role should they be assigned?. By default, the function assumes that the binary dummy variable columns created by the original variables will be used as predictors in a model.}

\item{trained}{A logical to indicate if the quantities for preprocessing have been estimated.}

\item{contrast}{A specification for which type of contrast should be used to make a set of full rank dummy variables. See \code{\link[stats]{contrasts}} for more details. \bold{not currently hooked up}}

\item{naming}{A function that defines the naming convention for new binary columns. See Details below.}

\item{levels}{A list that contains the information needed to create dummy variables for each variable contained in \code{terms}. This is \code{NULL} until the step is trained by \code{\link{learn.step_dummy}}.}

\item{x}{a \code{step_dummy} object that specifies which columns will be converted to dummy variables.}

\item{training}{A tibble or data frame that contains the training set. These data will be used to define the dummy variables for all future data when this step is applied.}

\item{info}{A tibble with information on the current set of columns in the design matrix.}

\item{...}{further arguments passed to or from other methods (not currently used).}

\item{object}{A trained step object.}

\item{newdata}{A tibble or data frame that has nominal variables that will be converted to dumy variables.}
}
\value{
\code{step_dummy} and \code{learn.step_dummy} return objects of class \code{step_dummy}.

\code{process.step_dummy} returns a tibble of processed data.
}
\description{
\code{step_dummy} creates a a \emph{specification} of a recipe step that will convert nominal data (e.g. character or factors) into one or more numeric binary model terms for the levels of the original data. For example, if a factor column in the data set has levels of "red", "green", "blue", the dummy variable process will create two additional columns of 0/1 data for two of those three values (and remove the original column).

For a training set of data, \code{learn.step_dummy} enumerates the possible values of the variables so that dummy variables can be created when a specific data set is \emph{processed}. This function is \emph{not} intended to be directly called by the user.

\code{process.step_dummy} is used to apply the process of creating dummy variables to any data set. This creates new columns in the data set and removes the original column(s). This function is \emph{not} intended to be directly called by the user.
}
\concept{
preprocessing dummy_variables model_specification dummy_variables variable_encodings
}
\keyword{datagen}

