% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/holiday.R
\name{step_holiday}
\alias{step_holiday}
\title{Holiday Feature Generator}
\usage{
step_holiday(recipe, ..., role = "predictor", trained = FALSE,
  holidays = c("LaborDay", "NewYearsDay", "ChristmasDay"), variables = NULL)
}
\arguments{
\item{recipe}{A recipe object. The step will be added to the sequence of 
operations for this recipe.}

\item{...}{One or more selector functions to choose which variables will be
used to create the new variables. The selected variables should have
class \code{Date} or \code{POSIXct}. See \code{\link{selections}} for
more details.}

\item{role}{For model terms created by this step, what analysis role should
they be assigned?. By default, the function assumes that the new variable
columns created by the original variables will be used as predictors in
a model.}

\item{trained}{A logical to indicate if the quantities for preprocessing 
have been estimated.}

\item{holidays}{A character string that includes at least one holdiay
supported by the \code{timeDate} package. See
\code{\link[timeDate]{listHolidays}} for a complete list.}

\item{variables}{A character string of variables that will be used as
inputs. This field is a placeholder and will be populated once
\code{\link{learn.recipe}} is used.}
}
\value{
\code{step_holiday} returns an object of class \code{step_holiday}.
}
\description{
\code{step_holiday} creates a a \emph{specification} of a recipe step that
  will convert date data into one or more binary indicator variables for
  common holidays.
}
\details{
Unlike other steps, \code{step_holiday} does \emph{not} remove the
  original date variables. \code{\link{step_rm}} can be used for
  this purpose.
}
\examples{
library(lubridate)

examples <- data.frame(someday = ymd("2000-12-20") + days(0:40))
holiday_rec <- recipe(~ someday, examples) \%>\%
   step_holiday(all_predictors())

holiday_rec <- learn(holiday_rec, training = examples)
holiday_values <- process(holiday_rec, newdata = examples)
holiday_values
}
\seealso{
\code{\link{step_date}} \code{\link{step_rm}}
  \code{\link{recipe}} \code{\link{learn.recipe}}
  \code{\link{process.recipe}} \code{\link[timeDate]{listHolidays}}
}
\concept{
preprocessing model_specification variable_encodings dates
}
\keyword{datagen}
