% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hyperbolic.R
\name{step_hyperbolic}
\alias{step_hyperbolic}
\title{Hyperbolic Transformations}
\usage{
step_hyperbolic(recipe, terms, role = NA, trained = FALSE, func = "sin",
  inverse = TRUE, vars = NULL)
}
\arguments{
\item{recipe}{A recipe object. The step will be added to the sequence of operations for this recipe.}

\item{terms}{A formula that represents the variables or terms that will be processed. The raw variable names can be used or \pkg{dplyr} selection tools. See \code{\link{selections}} for more details.}

\item{role}{Not used by this step since no new variables are created.}

\item{trained}{A logical to indicate if the quantities for preprocessing have been estimated.}

\item{func}{A character value for the function. Valid values are "sin", "cos", or "tan".}

\item{inverse}{A logical: should the inverse function be used?}

\item{vars}{A character string of variable names that will be (eventually) populated by the \code{terms} argument.}
}
\value{
\code{step_hyperbolic} returns an object of class \code{step_hyperbolic}.
}
\description{
\code{step_hyperbolic} creates a \emph{specification} of a recipe step that will transform data using a hyperbolic function.
}
\examples{
set.seed(313)
examples <- matrix(rnorm(40), ncol = 2)
examples <- as.data.frame(examples)

rec <- recipe(~ V1 + V2, data = examples)

library(magrittr)
cos_trans <- rec  \%>\%
  step_hyperbolic(terms = ~ predictors(),
                  fun = "cos", inverse = FALSE)

cos_obj <- learn(cos_trans, training = examples)

transformed_te <- process(cos_obj, examples)
plot(examples$V1, transformed_te$V1)
}
\seealso{
\code{\link{step_logit}} \code{\link{step_invlogit}} \code{\link{step_log}}  \code{\link{step_sqrt}}    \code{\link{recipe}} \code{\link{learn.recipe}} \code{\link{process.recipe}}
}
\concept{
preprocessing transformation_methods
}
\keyword{datagen}

