% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interactions.R
\name{step_interact}
\alias{step_interact}
\title{Declare Interaction Variables}
\usage{
step_interact(recipe, terms, role = "predictor", trained = FALSE,
  objects = NULL, sep = "_x_")
}
\arguments{
\item{recipe}{A recipe object. The step will be added to the sequence of operations for this recipe.}

\item{terms}{A traditional R formula that contains interaction terms.}

\item{role}{For model terms created by this step, what analysis role should they be assigned?. By default, the function assumes that the new columns created from the original variables will be used as predictors in a model.}

\item{trained}{A logical to indicate if the quantities for preprocessing have been estimated.}

\item{objects}{A list of \code{terms} objects for each individual interation.}

\item{sep}{A character value used to delinate variables in an interaction (e.g. \code{var1_x_var2} instead of the more traditional \code{var1:var2}.}
}
\value{
An object of class \code{interact_step}.
}
\description{
This function is a \emph{specification} of a recipe step that will create new columns that are interaction terms between two or more variables
}
\author{
Max Kuhn
}
\concept{
preprocessing
}
\keyword{datagen}

