% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interactions.R
\name{step_interact}
\alias{learn.interact_step}
\alias{process.interact_step}
\alias{step_interact}
\title{Create Interaction Variables}
\usage{
step_interact(recipe, terms, role = "predictor", trained = FALSE,
  objects = NULL, sep = "_x_")

\method{learn}{interact_step}(x, training, ...)

\method{process}{interact_step}(object, newdata, ...)
}
\arguments{
\item{recipe}{A recipe object. The step will be added to the sequence of operations for this recipe.}

\item{terms}{A traditional R formula that contains interaction terms.}

\item{role}{For model terms created by this step, what analysis role should they be assigned?. By default, the function assumes that the new columns created from the original variables will be used as predictors in a model.}

\item{trained}{A logical to indicate if the quantities for preprocessing have been estimated.}

\item{objects}{A list of \code{terms} objects for each individual interation.}

\item{sep}{A character value used to delinate variables in an interaction (e.g. \code{var1_x_var2} instead of the more traditional \code{var1:var2}.}

\item{x}{a \code{interact_step} object}

\item{training}{a tibble or data frame that contains the training set.}

\item{...}{further arguments passed to or from other methods (not currently used).}

\item{object}{A trained step object.}

\item{newdata}{A tibble or data frame that has numeric variables for the interactions.}
}
\value{
\code{step_interact} and \code{learn.interact_step} return objects of class \code{interact_step}.

\code{process.interact_step} returns a tibble of processed data.
}
\description{
\code{step_interact} creates a \emph{specification} of a recipe step that will create new columns that are interaction terms between two or more variables.

For a training set of data, \code{learn.interact_step} computes the required information to produce interaction terms. Note that no interactions are created by this function.

\code{process.interact_step} augment the current data with columns containing the interactions.
}
\concept{
preprocessing
}
\keyword{datagen}

