% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isomap.R
\name{step_isomap}
\alias{step_isomap}
\title{Isomap Embedding.}
\usage{
step_isomap(recipe, terms, role = "predictor", trained = FALSE, num = 5,
  options = list(knn = 50), res = NULL, prefix = "Isomap")
}
\arguments{
\item{recipe}{A recipe object. The step will be added to the sequence of operations for this recipe.}

\item{terms}{A representation of the variables or terms that will be used to compute the dimensions.}

\item{role}{For model terms created by this step, what analysis role should they be assigned?. By default, the function assumes that the new dimension columns created by the original variables will be used as predictors in a model.}

\item{trained}{A logical to indicate if the quantities for preprocessing have been estimated.}

\item{num}{The number of isomap dimensions to retain as new predictors. If \code{num} is greater than the number of columns or the number of possible dimensions, a smaller value will be used.}

\item{options}{A list of options to \code{\link[dimRed]{Isomap}}.}

\item{res}{The \code{\link[dimRed]{Isomap}} object is stored here once this preprocessing step has be trained by \code{\link{learn.recipe}}.}

\item{prefix}{A character string that will be the prefix to the resulting new variables. See notes below}
}
\value{
\code{step_isomap} returns an object of class \code{step_isomap}.
}
\description{
\code{step_isomap} creates a \emph{specification} of a recipe step that will convert numeric data into one or more new dimensions.
}
\concept{
preprocessing isomap projection_methods
}
\keyword{datagen}

