% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprec-impute_knn.R
\name{step_knnimpute}
\alias{step_knnimpute}
\title{Impute via k-nearest neighbors}
\usage{
step_knnimpute(
  recipe,
  ...,
  role = NA,
  trained = FALSE,
  neighbors = 5,
  impute_with = imp_vars(all_predictors()),
  options = list(nthread = 1, eps = 1e-08),
  ref_data = NULL,
  columns = NULL,
  skip = FALSE,
  id = rand_id("impute_knn")
)
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}

Please use \code{\link[=step_impute_knn]{step_impute_knn()}} instead.
}
\keyword{internal}
