% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kpca.R
\name{step_kpca}
\alias{learn.kpca_step}
\alias{process.kpca_step}
\alias{step_kpca}
\title{Kernel PCA Signal Extraction.}
\usage{
step_kpca(recipe, terms, role = "predictor", trained = FALSE, num = 5,
  res = NULL, options = list(kernel = "rbfdot", kpar = list(sigma = 0.2)))

\method{learn}{kpca_step}(x, training, ...)

\method{process}{kpca_step}(object, newdata, ...)
}
\arguments{
\item{recipe}{A recipe object. The step will be added to the sequence of operations for this recipe.}

\item{terms}{A representation of the variables or terms that will be used to compute the components.}

\item{role}{For model terms created by this step, what analysis role should they be assigned?. By default, the function assumes that the new principal component columns created by the original variables will be used as predictors in a model.}

\item{trained}{A logical to indicate if the quantities for preprocessing have been estimated.}

\item{num}{The number of PCA components to retain as new predictors. If \code{num} is greater than the number of columns or the number of possible components, a smaller value will be used.}

\item{res}{An S4 \code{\link[kernlab]{kpca}} object is stored here once this preprocessing step has be trained by \code{\link{learn.kpca_step}}.}

\item{options}{A list of options to \code{\link[kernlab]{kpca}}. Defaults are set for the arguments \code{kernel} and \code{kpar} but others can be passed in. \bold{Note} that the arguments \code{x} and \code{features} should not be passed here (or at all).}

\item{x}{A \code{kpca_step} object that contains the kernel PCA specifications.}

\item{training}{a tibble or data frame that contains the training set. These data will be used to compute the loadings that are used when this step is applied.}

\item{...}{further arguments passed to or from other methods (not currently used).}

\item{object}{A trained step object.}

\item{newdata}{A tibble or data frame that has numeric variables that will be processed.}
}
\value{
\code{step_kpca} and \code{learn.kpca_step} return objects of class \code{kpca_step}.

\code{process.kpca_step} returns a tibble of processed data.
}
\description{
\code{step_kpca} a \emph{specification} of a recipe step that will convert numeric data into one or more principal components using a kernel basis expansion.

For a training set of data, \code{learn.kpca_step} uses \code{\link[kernlab]{kpca}} to estimate the loadings for the principal components in the kernel space. This transformation only compute the required statistics for kernel PCA.

\code{process.kpca_step} is used to compute the components on specific data sets. This creates new columns in the data set and removes the original columns.
}
\concept{
preprocessing pca projection_methods kernel_methods
}
\keyword{datagen}

