% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lag.R
\name{step_lag}
\alias{step_lag}
\title{Lag a date valued predictor}
\usage{
step_lag(recipe, ..., role = "predictor", trained = FALSE, lag = 1,
  default = NA, prefix = "lag_")
}
\arguments{
\item{recipe}{A recipe object. The step will be added to the sequence of
operations for this recipe.}

\item{...}{One or more selector functions to choose which variables are
affected by the step. See \code{\link[=selections]{selections()}} for more details.}

\item{role}{Defaults to "predictor"}

\item{trained}{A logical to indicate if the quantities for preprocessing
have been estimated.}

\item{lag}{A vector of integers. Each specified column will be lagged be
each value in the vector.}

\item{default}{Passed to \code{dplyr::lag}, determines what fills empty rows
left by lagging (defaults to NA).}

\item{prefix}{A prefix for generated column names, default to "lag_".}
}
\value{
An updated version of \code{recipe} with the
new step added to the sequence of existing steps (if any).
}
\description{
\code{step_lag} creates a \emph{specification} of a recipe step that
will sort the data into chronological order for each selected
variable and then add a new column of lagged data. Lagged data will
by default include NA values where the lag was induced. These can be
removed with \code{\link[=step_naomit]{step_naomit()}}.
}
\examples{
n <- 10
start <- as.Date('1999/01/01')
end <- as.Date('2000/01/01')

df <- data.frame(x = runif(n), y = rnorm(n),
                 t = sample(seq(start, end, by = "day"), n))

baked <- recipe(~ ., data = df) \%>\%
  step_lag(has_type("Date"), lag = 2) \%>\%
  prep(df) \%>\%
  bake(df)

}
\seealso{
\code{\link[=recipe]{recipe()}} \code{\link[=prep.recipe]{prep.recipe()}} \code{\link[=bake.recipe]{bake.recipe()}} \code{\link[=step_naomit]{step_naomit()}}
}
