% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log.R
\name{step_log}
\alias{step_log}
\title{Logarithmic Transformation}
\usage{
step_log(recipe, terms, role = NA, trained = FALSE, base = exp(1),
  vars = NULL)
}
\arguments{
\item{recipe}{A recipe object. The step will be added to the sequence of operations for this recipe.}

\item{terms}{A formula that represents the variables or terms that will be processed. The raw variable names can be used or \pkg{dplyr} selection tools. See \code{\link{selections}} for more details.}

\item{role}{Not used by this step since no new variables are created.}

\item{trained}{A logical to indicate if the quantities for preprocessing have been estimated.}

\item{base}{A numeric value for the base.}

\item{vars}{A character string of variable names that will be (eventually) populated by the \code{terms} argument.}
}
\value{
\code{step_log} returns an object of class \code{step_log}.
}
\description{
\code{step_log} creates a \emph{specification} of a recipe step that will log transform data.
}
\examples{
set.seed(313)
examples <- matrix(exp(rnorm(40)), ncol = 2)
examples <- as.data.frame(examples)

rec <- recipe(~ V1 + V2, data = examples)

library(magrittr)
log_trans <- rec  \%>\%
  step_log(terms = ~ is_predictor())

log_obj <- learn(log_trans, training = examples)

transformed_te <- process(log_obj, examples)
plot(examples$V1, transformed_te$V1)
}
\concept{
preprocessing transformation_methods
}
\keyword{datagen}

