% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log.R
\name{step_log}
\alias{learn.step_log}
\alias{process.step_log}
\alias{step_log}
\title{Logarithmic Transformation}
\usage{
step_log(recipe, terms, role = NA, trained = FALSE, base = exp(1))

\method{learn}{step_log}(x, training, ...)

\method{process}{step_log}(object, newdata, ...)
}
\arguments{
\item{recipe}{A recipe object. The step will be added to the sequence of operations for this recipe.}

\item{terms}{A representation of the variables or terms that will be centered.}

\item{role}{Not used by this step since no new variables are created.}

\item{trained}{A logical to indicate if the quantities for preprocessing have been estimated.}

\item{base}{A numeric value for the base.}

\item{x}{a \code{step_log} object that specifies which columns will be transformed}

\item{training}{A tibble or data frame that contains the training set.}

\item{...}{further arguments passed to or from other methods (not currently used).}

\item{object}{A trained step object.}

\item{newdata}{A tibble or data frame that has numeric variables that will be transformed}
}
\value{
\code{step_log} and \code{learn.step_log} return objects of class \code{step_log}.

\code{process.step_log} returns a tibble of processed data.
}
\description{
\code{step_log} creates a \emph{specification} of a recipe step that will log transform data.

For a training set of data, \code{learn.step_log} configures the log transformation (by basically doing nothing).

\code{process.step_log} is used to transform columns on specific data sets. This replaces values in the original columns.
}
\concept{
preprocessing transformation_methods
}
\keyword{datagen}

