% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logit.R
\name{step_logit}
\alias{learn.step_logit}
\alias{process.step_logit}
\alias{step_logit}
\title{Logit Transformation}
\usage{
step_logit(recipe, terms, role = NA, trained = FALSE)

\method{learn}{step_logit}(x, training, ...)

\method{process}{step_logit}(object, newdata, ...)
}
\arguments{
\item{recipe}{A recipe object. The step will be added to the sequence of operations for this recipe.}

\item{terms}{A representation of the variables or terms that will be transformed.}

\item{role}{Not used by this step since no new variables are created.}

\item{trained}{A logical to indicate if the quantities for preprocessing have been estimated.}

\item{x}{a \code{step_logit} object that specifies which columns will be transformed}

\item{training}{a tibble or data frame that contains the training set.}

\item{...}{further arguments passed to or from other methods (not currently used).}

\item{object}{A trained step object.}

\item{newdata}{A tibble or data frame that has numeric variables that will be transformed}
}
\value{
\code{step_logit} and \code{learn.step_logit} return objects of class \code{step_logit}.

\code{process.step_logit} returns a tibble of processed data.
}
\description{
\code{step_logit} creates a \emph{specification} of a recipe step that will logit transform the data.

For a training set of data, \code{learn.step_logit} configures the logit transformation (by basically doing nothing).

\code{process.step_logit} is used to transform columns on specific data sets. This replaces values in the original columns.
}
\concept{
preprocessing transformation_methods
}
\keyword{datagen}

