% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprec-impute_lower.R
\name{step_lowerimpute}
\alias{step_lowerimpute}
\title{Impute numeric data below the threshold of measurement}
\usage{
step_lowerimpute(
  recipe,
  ...,
  role = NA,
  trained = FALSE,
  threshold = NULL,
  skip = FALSE,
  id = rand_id("impute_lower")
)
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}

Please use \code{\link[=step_impute_lower]{step_impute_lower()}} instead.
}
\keyword{internal}
