% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprec-impute_mena.R
\name{step_meanimpute}
\alias{step_meanimpute}
\title{Impute numeric data using the mean}
\usage{
step_meanimpute(
  recipe,
  ...,
  role = NA,
  trained = FALSE,
  means = NULL,
  trim = 0,
  skip = FALSE,
  id = rand_id("impute_mean")
)
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}

Please use \code{\link[=step_impute_mean]{step_impute_mean()}} instead.
}
\keyword{internal}
