% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprec-impute_median.R
\name{step_medianimpute}
\alias{step_medianimpute}
\title{Impute numeric data using the median}
\usage{
step_medianimpute(
  recipe,
  ...,
  role = NA,
  trained = FALSE,
  medians = NULL,
  skip = FALSE,
  id = rand_id("impute_median")
)
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}

Please use \code{\link[=step_impute_median]{step_impute_median()}} instead.
}
\keyword{internal}
