% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprec-impute_mode.R
\name{step_modeimpute}
\alias{step_modeimpute}
\title{Impute nominal data using the most common value}
\usage{
step_modeimpute(
  recipe,
  ...,
  role = NA,
  trained = FALSE,
  modes = NULL,
  ptype = NULL,
  skip = FALSE,
  id = rand_id("impute_mode")
)
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}

Please use \code{\link[=step_impute_mode]{step_impute_mode()}} instead.
}
\keyword{internal}
