% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nzv.R
\name{step_nzv}
\alias{step_nzv}
\title{Declare Which Variables Are Used Near-Zero Variance Filter.}
\usage{
step_nzv(recipe, terms, role = NA)
}
\arguments{
\item{recipe}{A recipe object. The step will be added to the sequence of operations for this recipe.}

\item{terms}{A representation of the variables or terms that will evaluated by the filtering process.}

\item{role}{Not used by this step since no new variables are created.}
}
\value{
An object of class \code{nzv_step}.
}
\description{
This function is a \emph{specification} of a recipe step that will potentially remove variables that are highly sparse and unbalanced.
}
\author{
Max Kuhn
}
\concept{
preprocessing variable_filters
}
\keyword{datagen}

