% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nzv.R
\name{step_nzv}
\alias{learn.nzv_step}
\alias{process.nzv_step}
\alias{step_nzv}
\title{Near-Zero Variance Filter.}
\usage{
step_nzv(recipe, terms, role = NA, trained = FALSE, options = list(freqCut
  = 95/5, uniqueCut = 10), removals = NULL)

\method{learn}{nzv_step}(x, training, ...)

\method{process}{nzv_step}(object, newdata, ...)
}
\arguments{
\item{recipe}{A recipe object. The step will be added to the sequence of operations for this recipe.}

\item{terms}{A representation of the variables or terms that will evaluated by the filtering process.}

\item{role}{Not used by this step since no new variables are created.}

\item{trained}{A logical to indicate if the quantities for preprocessing have been estimated.}

\item{options}{A list of options for \code{\link[caret]{nearZeroVar}}. \bold{Note} that the arguments \code{data} and \code{names} should not be included in this list.}

\item{removals}{A character string that contains the names of columns that should be removed. These values are not determined until \code{\link{learn.nzv_step}}.}

\item{x}{a \code{nzv_step} object that contains the list of predictors that should be removed.}

\item{training}{a tibble or data frame that contains the training set.}

\item{...}{further arguments passed to or from other methods (not currently used).}

\item{object}{A trained step object.}

\item{newdata}{A tibble or data frame to be filtered.}
}
\value{
\code{step_nzv} and \code{learn.nzv_step} return objects of class \code{nzv_step}.

\code{process.nzv_step} returns a tibble of processed data.
}
\description{
\code{step_nzv} creates a \emph{specification} of a recipe step that will potentially remove variables that are highly sparse and unbalanced.

For a training set of data, \code{learn.nzv_step} determines which, if any, columns in the training set have sparse and unbalanced distributions.

\code{process.nzv_step} is used to potentially remove columns from the data.
}
\concept{
preprocessing variable_filters
}
\keyword{datagen}

