% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pca.R
\name{step_pca}
\alias{step_pca}
\title{Declare Which Variables Are Used for PCA Signal Extraction.}
\usage{
step_pca(recipe, terms, role = "predictor")
}
\arguments{
\item{recipe}{A recipe object. The step will be added to the sequence of operations for this recipe.}

\item{terms}{A representation of the variables or terms that will be used to compute the components.}

\item{role}{For model terms created by this step, what analysis role should they be assigned?. By default, the function assumes that the new principal component columns created by the original variables will be used as predictors in a model.}
}
\value{
An object of class \code{pca_step}.
}
\description{
This function is a \emph{specification} of a recipe step that will convert numeric data into one or more principal components.
}
\author{
Max Kuhn
}
\concept{
preprocessing pca projection_methods
}
\keyword{datagen}

