% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pca.R
\name{step_pca}
\alias{learn.pca_step}
\alias{process.pca_step}
\alias{step_pca}
\title{PCA Signal Extraction.}
\usage{
step_pca(recipe, terms, role = "predictor", trained = FALSE, num = 5,
  options = list(center = TRUE, scale. = TRUE, retx = FALSE), res = NULL)

\method{learn}{pca_step}(x, training, ...)

\method{process}{pca_step}(object, newdata, ...)
}
\arguments{
\item{recipe}{A recipe object. The step will be added to the sequence of operations for this recipe.}

\item{terms}{A representation of the variables or terms that will be used to compute the components.}

\item{role}{For model terms created by this step, what analysis role should they be assigned?. By default, the function assumes that the new principal component columns created by the original variables will be used as predictors in a model.}

\item{trained}{A logical to indicate if the quantities for preprocessing have been estimated.}

\item{num}{The number of PCA components to retain as new predictors. If \code{num} is greater than the number of columns or the number of possible components, a smaller value will be used.}

\item{options}{A list of options to \code{\link[stats]{prcomp}}. Defaults are set for the arguments \code{center} and \code{scale.} but others can be passed in (e.g. \code{tol}). \bold{Note} that the arguments \code{x} and \code{y} should not be passed here (or at all).}

\item{res}{The \code{\link[stats]{prcomp}} object is stored here once this preprocessing step has be trained by \code{\link{learn.pca_step}}.}

\item{x}{A \code{pca_step} object that contains the PCA specifications.}

\item{training}{a tibble or data frame that contains the training set. These data will be used to compute the loadings that are used when this step is applied.}

\item{...}{further arguments passed to or from other methods (not currently used).}

\item{object}{A trained step object.}

\item{newdata}{A tibble or data frame that has numeric variables that will be converted to principal components.}
}
\value{
\code{step_pca} and \code{learn.pca_step} return objects of class \code{pca_step}.

\code{process.pca_step} returns a tibble of processed data.
}
\description{
\code{step_pca} creates a \emph{specification} of a recipe step that will convert numeric data into one or more principal components.

For a training set of data, \code{learn.pca_step} estimates the loadings for the principal components. This transformation only compute the required statistics for PCA.

\code{process.pca_step} is used to compute the components on specific data sets. This creates new columns in the data set and removes the original columns.
}
\concept{
preprocessing pca projection_methods
}
\keyword{datagen}

