% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprec-impute_roll.R
\name{step_rollimpute}
\alias{step_rollimpute}
\title{Impute numeric data using a rolling window statistic}
\usage{
step_rollimpute(
  recipe,
  ...,
  role = NA,
  trained = FALSE,
  columns = NULL,
  statistic = median,
  window = 5,
  skip = FALSE,
  id = rand_id("impute_roll")
)
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}

Please use \code{\link[=step_impute_roll]{step_impute_roll()}} instead.
}
\keyword{internal}
