% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale.R
\name{step_scale}
\alias{step_scale}
\title{Declare Variables for Scaling}
\usage{
step_scale(recipe, terms, role = NA, sds = NULL)
}
\arguments{
\item{recipe}{A recipe object. The step will be added to the sequence of operations for this recipe.}

\item{terms}{A representation of the variables or terms that will be scaled.}

\item{role}{Not used by this step since no new variables are created.}

\item{sds}{A named numeric vector of standard deviations This is \code{NULL} until computed by \code{\link{learn.scale_step}}.}
}
\value{
An object of class \code{scale_step}.
}
\description{
This function is a \emph{specification} of a recipe step that will normalize numeric data to have a standard deviation of one
}
\author{
Max Kuhn
}
\concept{
preprocessing normalization_methods
}
\keyword{datagen}

