% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale.R
\name{step_scale}
\alias{learn.scale_step}
\alias{process.scale_step}
\alias{step_scale}
\title{Scaling Numeric Data}
\usage{
step_scale(recipe, terms, role = NA, trained = FALSE, sds = NULL)

\method{learn}{scale_step}(x, data, ...)

\method{process}{scale_step}(x, data, ...)
}
\arguments{
\item{recipe}{A recipe object. The step will be added to the sequence of operations for this recipe.}

\item{terms}{A representation of the variables or terms that will be scaled.}

\item{role}{Not used by this step since no new variables are created.}

\item{trained}{A logical to indicate if the quantities for preprocessing have been estimated.}

\item{sds}{A named numeric vector of standard deviations This is \code{NULL} until computed by \code{\link{learn.scale_step}}.}

\item{x}{a \code{scale_step} object that specifies which columns will be scaled}

\item{data}{a tibble or data frame that contains the training set.}

\item{...}{further arguments passed to or from other methods (not currently used).}

\item{data}{A tibble or data frame that has numeric variables that will be scaled}
}
\value{
\code{step_scale} and \code{learn.scale_step} return objects of class \code{scale_step}.

A tibble of processed data.
}
\description{
\code{step_scale} creates a \emph{specification} of a recipe step that will normalize numeric data to have a standard deviation of one.

For a training set of data, \code{learn.scale_step} estimates the standard deviations from numeric columns.

\code{process.scale_step} is used to perform the scaling on specific data sets. This replaces values in the original columns.
}
\concept{
preprocessing normalization_methods
}
\keyword{datagen}

