% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale.R
\name{step_scale}
\alias{step_scale}
\title{Scaling Numeric Data}
\usage{
step_scale(recipe, terms, role = NA, trained = FALSE, sds = NULL)
}
\arguments{
\item{recipe}{A recipe object. The step will be added to the sequence of operations for this recipe.}

\item{terms}{A representation of the variables or terms that will be scaled.}

\item{role}{Not used by this step since no new variables are created.}

\item{trained}{A logical to indicate if the quantities for preprocessing have been estimated.}

\item{sds}{A named numeric vector of standard deviations This is \code{NULL} until computed by \code{\link{learn.recipe}}.}
}
\value{
\code{step_scale}  returns an object of class \code{step_scale}.
}
\description{
\code{step_scale} creates a \emph{specification} of a recipe step that will normalize numeric data to have a standard deviation of one.
}
\concept{
preprocessing normalization_methods
}
\keyword{datagen}

