% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale.R
\name{step_scale}
\alias{learn.step_scale}
\alias{process.step_scale}
\alias{step_scale}
\title{Scaling Numeric Data}
\usage{
step_scale(recipe, terms, role = NA, trained = FALSE, sds = NULL)

\method{learn}{step_scale}(x, training, info = NULL, ...)

\method{process}{step_scale}(object, newdata, ...)
}
\arguments{
\item{recipe}{A recipe object. The step will be added to the sequence of operations for this recipe.}

\item{terms}{A representation of the variables or terms that will be scaled.}

\item{role}{Not used by this step since no new variables are created.}

\item{trained}{A logical to indicate if the quantities for preprocessing have been estimated.}

\item{sds}{A named numeric vector of standard deviations This is \code{NULL} until computed by \code{\link{learn.step_scale}}.}

\item{x}{a \code{step_scale} object that specifies which columns will be scaled}

\item{training}{A tibble or data frame that contains the training set.}

\item{info}{A tibble with information on the current set of columns in the design matrix.}

\item{...}{further arguments passed to or from other methods (not currently used).}

\item{object}{A trained step object.}

\item{newdata}{A tibble or data frame that has numeric variables that will be scaled}
}
\value{
\code{step_scale} and \code{learn.step_scale} return objects of class \code{step_scale}.

\code{process.step_scale} returns a tibble of processed data.
}
\description{
\code{step_scale} creates a \emph{specification} of a recipe step that will normalize numeric data to have a standard deviation of one.

For a training set of data, \code{learn.step_scale} estimates the standard deviations from numeric columns. This function is \emph{not} intended to be directly called by the user.

\code{process.step_scale} is used to perform the scaling on specific data sets. This replaces values in the original columns. This function is \emph{not} intended to be directly called by the user.
}
\concept{
preprocessing normalization_methods
}
\keyword{datagen}

