% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatialsign.R
\name{step_spatialsign}
\alias{learn.spatialsign_step}
\alias{process.spatialsign_step}
\alias{step_spatialsign}
\title{Declare Which Variables for the Spatial Sign.}
\usage{
step_spatialsign(recipe, terms, role = "predictor", trained = FALSE)

\method{learn}{spatialsign_step}(x, data, ...)

\method{process}{spatialsign_step}(x, data, ...)
}
\arguments{
\item{recipe}{A recipe object. The step will be added to the sequence of operations for this recipe.}

\item{terms}{A representation of the variables or terms that will be used for the normalization.}

\item{role}{For model terms created by this step, what analysis role should they be assigned?}

\item{trained}{A logical to indicate if the quantities for preprocessing have been estimated.}

\item{x}{a \code{spatialsign_step} object that contains the spatial sign objects.}

\item{data}{a tibble or data frame that contains the training set. These data will be used to compute the loadings that are used when this step is applied.}

\item{...}{further arguments passed to or from other methods (not currently used).}

\item{x}{A trained \code{spatialsign_step} object.}

\item{data}{A tibble or data frame that has numeric variables that will be transformed.}

\item{...}{further arguments passed to or from other methods (not currently used).}
}
\value{
An object of class \code{spatialsign_step}.

An object of class \code{spatialsign_step}.

A tibble of processed data.
}
\description{
This function is a \emph{specification} of a recipe step that will convert numeric data into a projection on to a unit sphere.

The spatial sign is a transformation of the original variables that does not require any parameters to be estimated. As such, this function only returns the original object.

For a \code{spatialsign_step} object, this function projects the current data into a unit sphere. This creates overwrites the original columns
}
\concept{
preprocessing projection_methods

preprocessing projection_methods

preprocessing projection_methods
}
\keyword{datagen}

