% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatialsign.R
\name{step_spatialsign}
\alias{step_spatialsign}
\title{Declare Which Variables for the Spatial Sign.}
\usage{
step_spatialsign(recipe, terms, role = "predictor", trained = FALSE)
}
\arguments{
\item{recipe}{A recipe object. The step will be added to the sequence of operations for this recipe.}

\item{terms}{A representation of the variables or terms that will be used for the normalization.}

\item{role}{For model terms created by this step, what analysis role should they be assigned?}

\item{trained}{A logical to indicate if the quantities for preprocessing have been estimated.}
}
\value{
An object of class \code{spatialsign_step}.
}
\description{
This function is a \emph{specification} of a recipe step that will convert numeric data into a projection on to a unit sphere.
}
\author{
Max Kuhn
}
\concept{
preprocessing projection_methods
}
\keyword{datagen}

