% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatialsign.R
\name{step_spatialsign}
\alias{learn.step_spatialsign}
\alias{process.step_spatialsign}
\alias{step_spatialsign}
\title{Spatial Sign Preprocessing.}
\usage{
step_spatialsign(recipe, terms, role = "predictor", trained = FALSE)

\method{learn}{step_spatialsign}(x, training, ...)

\method{process}{step_spatialsign}(object, newdata, ...)
}
\arguments{
\item{recipe}{A recipe object. The step will be added to the sequence of operations for this recipe.}

\item{terms}{A representation of the variables or terms that will be used for the normalization.}

\item{role}{For model terms created by this step, what analysis role should they be assigned?}

\item{trained}{A logical to indicate if the quantities for preprocessing have been estimated.}

\item{x}{a \code{step_spatialsign} object that contains the spatial sign objects.}

\item{training}{a tibble or data frame that contains the training set. These data will be used to compute the loadings that are used when this step is applied.}

\item{...}{further arguments passed to or from other methods (not currently used).}

\item{object}{A trained step object.}

\item{newdata}{A tibble or data frame that has numeric variables that will be processed.}
}
\value{
\code{step_spatialsign} and \code{learn.step_spatialsign} return objects of class \code{step_spatialsign}.

\code{process.step_spatialsign} returns a tibble of processed data.
}
\description{
\code{step_spatialsign} is a \emph{specification} of a recipe step that will convert numeric data into a projection on to a unit sphere.

\code{learn.step_spatialsign} processes a training data set for the transformation.

To project the data onto a unit sphere, \code{process.step_spatialsign} is used and this overwrites the original columns.
}
\concept{
preprocessing projection_methods
}
\keyword{datagen}

