% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sqrt.R
\name{step_sqrt}
\alias{step_sqrt}
\title{Square Root Transformation}
\usage{
step_sqrt(recipe, terms, role = NA, trained = FALSE, vars = NULL)
}
\arguments{
\item{recipe}{A recipe object. The step will be added to the sequence of operations for this recipe.}

\item{terms}{A representation of the variables or terms that will be transformed.}

\item{role}{Not used by this step since no new variables are created.}

\item{trained}{A logical to indicate if the quantities for preprocessing have been estimated.}

\item{vars}{A character string of variable names that will be (eventually) populated by the \code{terms} argument.}
}
\value{
\code{step_sqrt}  returns an object of class \code{step_sqrt}.
}
\description{
\code{step_sqrt} creates a \emph{specification} of a recipe step that will square root transform the data.
}
\concept{
preprocessing transformation_methods
}
\keyword{datagen}

