% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/YeoJohnson.R
\name{step_YeoJohnson}
\alias{step_YeoJohnson}
\alias{tidy.step_YeoJohnson}
\title{Yeo-Johnson Transformation}
\usage{
step_YeoJohnson(recipe, ..., role = NA, trained = FALSE,
  lambdas = NULL, limits = c(-5, 5), nunique = 5, na.rm = TRUE,
  skip = FALSE)

\method{tidy}{step_YeoJohnson}(x, ...)
}
\arguments{
\item{recipe}{A recipe object. The step will be added to the
sequence of operations for this recipe.}

\item{...}{One or more selector functions to choose which
variables are affected by the step. See \code{\link[=selections]{selections()}}
for more details. For the \code{tidy} method, these are not
currently used.}

\item{role}{Not used by this step since no new variables are
created.}

\item{trained}{A logical to indicate if the quantities for
preprocessing have been estimated.}

\item{lambdas}{A numeric vector of transformation values. This
is \code{NULL} until computed by \code{\link[=prep.recipe]{prep.recipe()}}.}

\item{limits}{A length 2 numeric vector defining the range to
compute the transformation parameter lambda.}

\item{nunique}{An integer where data that have less possible
values will not be evaluate for a transformation.}

\item{na.rm}{A logical value indicating whether \code{NA}
values should be removed during computations.}

\item{skip}{A logical. Should the step be skipped when the
recipe is baked by \code{\link[=bake.recipe]{bake.recipe()}}? While all operations are baked
when \code{\link[=prep.recipe]{prep.recipe()}} is run, some operations may not be able to be
conducted on new data (e.g. processing the outcome variable(s)).
Care should be taken when using \code{skip = TRUE} as it may affect
the computations for subsequent operations}

\item{x}{A \code{step_YeoJohnson} object.}
}
\value{
An updated version of \code{recipe} with the new step
added to the sequence of existing steps (if any). For the
\code{tidy} method, a tibble with columns \code{terms} (the
selectors or variables selected) and \code{value} (the
lambda estimate).
}
\description{
\code{step_YeoJohnson} creates a \emph{specification} of a
recipe step that will transform data using a simple Yeo-Johnson
transformation.
}
\details{
The Yeo-Johnson transformation is very similar to the
Box-Cox but does not require the input variables to be strictly
positive. In the package, the partial log-likelihood function is
directly optimized within a reasonable set of transformation
values (which can be changed by the user).

This transformation is typically done on the outcome variable
using the residuals for a statistical model (such as ordinary
least squares). Here, a simple null model (intercept only) is
used to apply the transformation to the \emph{predictor}
variables individually. This can have the effect of making the
variable distributions more symmetric.

If the transformation parameters are estimated to be very
closed to the bounds, or if the optimization fails, a value of
\code{NA} is used and no transformation is applied.
}
\examples{

data(biomass)

biomass_tr <- biomass[biomass$dataset == "Training",]
biomass_te <- biomass[biomass$dataset == "Testing",]

rec <- recipe(HHV ~ carbon + hydrogen + oxygen + nitrogen + sulfur,
              data = biomass_tr)

yj_trans <- step_YeoJohnson(rec,  all_numeric())

yj_estimates <- prep(yj_trans, training = biomass_tr)

yj_te <- bake(yj_estimates, biomass_te)

plot(density(biomass_te$sulfur), main = "before")
plot(density(yj_te$sulfur), main = "after")

tidy(yj_trans, number = 1)
tidy(yj_estimates, number = 1)
}
\references{
Yeo, I. K., and Johnson, R. A. (2000). A new family of power
transformations to improve normality or symmetry. \emph{Biometrika}.
}
\seealso{
\code{\link[=step_BoxCox]{step_BoxCox()}} \code{\link[=recipe]{recipe()}}
\code{\link[=prep.recipe]{prep.recipe()}} \code{\link[=bake.recipe]{bake.recipe()}}
}
\concept{preprocessing transformation_methods}
\keyword{datagen}
