% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprec-terms_select.R
\name{terms_select}
\alias{terms_select}
\title{Select terms in a step function.}
\usage{
terms_select(terms, info, empty_fun = function(x) x)
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}

Please use \code{\link[=recipes_eval_select]{recipes_eval_select()}} instead.
}
\keyword{internal}
