% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bs.R, R/corr.R, R/discretize.R, R/downsample.R,
%   R/harmonic.R, R/ica.R, R/impute_bag.R, R/impute_knn.R, R/impute_mean.R,
%   R/impute_roll.R, R/isomap.R, R/kpca_poly.R, R/kpca_rbf.R, R/nnmf.R,
%   R/nnmf_sparse.R, R/ns.R, R/nzv.R, R/other.R, R/pca.R, R/pls.R, R/poly.R,
%   R/upsample.R, R/window.R
\name{tunable.step}
\alias{tunable.step}
\alias{tunable.step_bs}
\alias{tunable.step_corr}
\alias{tunable.step_discretize}
\alias{tunable.step_downsample}
\alias{tunable.step_harmonic}
\alias{tunable.step_ica}
\alias{tunable.step_impute_bag}
\alias{tunable.step_impute_knn}
\alias{tunable.step_impute_mean}
\alias{tunable.step_impute_roll}
\alias{tunable.step_isomap}
\alias{tunable.step_kpca_poly}
\alias{tunable.step_kpca_rbf}
\alias{tunable.step_nnmf}
\alias{tunable.step_nnmf_sparse}
\alias{tunable.step_ns}
\alias{tunable.step_nzv}
\alias{tunable.step_other}
\alias{tunable.step_pca}
\alias{tunable.step_pls}
\alias{tunable.step_poly}
\alias{tunable.step_upsample}
\alias{tunable.step_window}
\title{Information on what step arguments can be tuned}
\usage{
\method{tunable}{step_bs}(x, ...)

\method{tunable}{step_corr}(x, ...)

\method{tunable}{step_discretize}(x, ...)

\method{tunable}{step_downsample}(x, ...)

\method{tunable}{step_harmonic}(x, ...)

\method{tunable}{step_ica}(x, ...)

\method{tunable}{step_impute_bag}(x, ...)

\method{tunable}{step_impute_knn}(x, ...)

\method{tunable}{step_impute_mean}(x, ...)

\method{tunable}{step_impute_roll}(x, ...)

\method{tunable}{step_isomap}(x, ...)

\method{tunable}{step_kpca_poly}(x, ...)

\method{tunable}{step_kpca_rbf}(x, ...)

\method{tunable}{step_nnmf}(x, ...)

\method{tunable}{step_nnmf_sparse}(x, ...)

\method{tunable}{step_ns}(x, ...)

\method{tunable}{step_nzv}(x, ...)

\method{tunable}{step_other}(x, ...)

\method{tunable}{step_pca}(x, ...)

\method{tunable}{step_pls}(x, ...)

\method{tunable}{step_poly}(x, ...)

\method{tunable}{step_upsample}(x, ...)

\method{tunable}{step_window}(x, ...)
}
\description{
Information on what step arguments can be tuned
}
