% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/woe.R
\name{woe_table}
\alias{woe_table}
\title{Crosstable with woe between a binary outcome and a predictor variable.}
\usage{
woe_table(predictor, outcome, Laplace = 1e-06)
}
\arguments{
\item{predictor}{A atomic vector, usualy with few distinct values.}

\item{outcome}{The dependent variable. A atomic vector with exactly 2 distinct values.}

\item{Laplace}{Default to 1e-6. The \code{pseudocount} parameter of the Laplace Smoothing
estimator. Value to avoid -Inf/Inf from predictor category with only one outcome class.
Set to 0 to allow Inf/-Inf.}
}
\value{
a tibble with counts, proportions and woe.
Warning: woe can possibly be -Inf. Use 'Laplace' arg to avoid that.
}
\description{
Calculates some summaries and the WoE (Weight of Evidence) between a binary
outcome and a given predictor variable. Used to biuld the dictionary.
}
\examples{

outc <- rep(c("A", "B"), 20)
pred <- sample(c("X", "Y", "Z", "W"), size = 40, replace = TRUE)
woe_table(pred, outc)

# offset avoid Inf/-Inf
woe_table(c("A", "A", "B", "B"), c(0, 0, 0, 1), Laplace = 1e-6)
woe_table(c("A", "A", "B", "B"), c(0, 0, 0, 1), Laplace = 0)

}
\references{
Kullback, S. (1959). \emph{Information Theory and Statistics.} Wiley, New York.

Hastie, T., Tibshirani, R. and Friedman, J. (1986). \emph{Elements of Statistical Learning}, Second Edition, Springer, 2009.
}
